Cover rest of the method of the AdobeStockClient module Client class with stub functionality
