+++ AdobeStockClientStub/Model/DataProvider/RawStockResponse.php
@@ -1,0 +1,469 @@
+<?php
+/**
+ * Copyright © Magento, Inc. All rights reserved.
+ * See COPYING.txt for license details.
+ */
+declare(strict_types=1);
+
+namespace Magento\AdobeStockClientStub\Model\DataProvider;
+
+/**
+ * Represent the stub raw stock data response from the Adobe Stock service
+ */
+class RawStockResponse
+{
+    /**
+     * Adobe Stock service raw response
+     *
+     * @return array
+     */
+    public function getRawAdobeStockResponse(): array
+    {
+        return [
+            'nb_results' => 3,
+            'files' => [
+                0 => [
+                    'id' => 244680548,
+                    'title' => 'Cheerful senior woman on a swing at a playground',
+                    'creator_id' => 204567087,
+                    'creator_name' => 'Rawpixel.com',
+                    'creation_date' => '2019-01-22 10:18:16',
+                    'country_name' => 'United Kingdom of Great Britain and Northern Ireland',
+                    'thumbnail_url' => null,
+                    'thumbnail_html_tag' => null,
+                    'thumbnail_width' => null,
+                    'thumbnail_height' => null,
+                    'thumbnail_110_url' => null,
+                    'thumbnail_110_width' => null,
+                    'thumbnail_110_height' => null,
+                    'thumbnail_160_url' => null,
+                    'thumbnail_160_width' => null,
+                    'thumbnail_160_height' => null,
+                    'thumbnail_220_url' => null,
+                    'thumbnail_220_width' => null,
+                    'thumbnail_220_height' => null,
+                    'thumbnail_240_url' => 'http://image_url.jpg',
+                    'thumbnail_240_width' => null,
+                    'thumbnail_240_height' => null,
+                    'thumbnail_500_url' => 'http://image_url.jpg',
+                    'thumbnail_500_width' => null,
+                    'thumbnail_500_height' => null,
+                    'thumbnail_1000_url' => null,
+                    'thumbnail_1000_width' => null,
+                    'thumbnail_1000_height' => null,
+                    'media_type_id' => 1,
+                    'width' => 5040,
+                    'height' => 3360,
+                    'is_licensed' => null,
+                    'comp_url' => 'https://stock.adobe.com/Rest/Libraries/Watermarked/Download/244680548/2',
+                    'comp_width' => null,
+                    'comp_height' => null,
+                    'nb_views' => null,
+                    'nb_downloads' => null,
+                    'category' => [
+                        'id' => 695,
+                        'name' => 'People',
+                        'link' => null,
+                    ],
+                    'keywords' => [
+                        0 => [
+                            'name' => 'woman',
+                        ],
+                        1 => [
+                            'name' => 'adult',
+                        ],
+                        2 => [
+                            'name' => 'senior',
+                        ],
+                        3 => [
+                            'name' => 'caucasian',
+                        ],
+                        4 => [
+                            'name' => 'alone',
+                        ],
+                        5 => [
+                            'name' => 'swing',
+                        ],
+                        6 => [
+                            'name' => 'playground',
+                        ],
+                        7 => [
+                            'name' => 'smile',
+                        ],
+                        8 => [
+                            'name' => 'white hair',
+                        ],
+                        9 => [
+                            'name' => 'shirt',
+                        ],
+                        10 => [
+                            'name' => 'chain',
+                        ],
+                        11 => [
+                            'name' => 'park',
+                        ],
+                        12 => [
+                            'name' => 'fun',
+                        ],
+                        13 => [
+                            'name' => 'cheerful',
+                        ],
+                        14 => [
+                            'name' => 'vitality',
+                        ],
+                        15 => [
+                            'name' => 'retired',
+                        ],
+                        16 => [
+                            'name' => 'happiness',
+                        ],
+                        17 => [
+                            'name' => 'enjoyment',
+                        ],
+                        18 => [
+                            'name' => 'outdoors',
+                        ],
+                        19 => [
+                            'name' => 'day',
+                        ],
+                        20 => [
+                            'name' => '1',
+                        ],
+                        21 => [
+                            'name' => 'people',
+                        ],
+                    ],
+                    'has_releases' => null,
+                    'asset_type_id' => null,
+                    'vector_type' => null,
+                    'content_type' => 'image/jpeg',
+                    'framerate' => null,
+                    'duration' => null,
+                    'stock_id' => null,
+                    'comps' => null,
+                    'details_url' => 'http://image_url.jpg',
+                    'template_type_id' => null,
+                    'template_category_ids' => null,
+                    'marketing_text' => null,
+                    'description' => null,
+                    'size_bytes' => null,
+                    'premium_level_id' => 0,
+                    'is_premium' => null,
+                    'licenses' => null,
+                    'video_preview_url' => null,
+                    'VideoPreviewHeight' => null,
+                    'video_preview_width' => null,
+                    'video_preview_content_length' => null,
+                    'video_preview_content_type' => null,
+                    'video_small_preview_url' => null,
+                    'video_small_preview_height' => null,
+                    'video_small_preview_width' => null,
+                    'video_small_preview_content_length' => null,
+                    'video_small_preview_content_type' => null,
+                ],
+                1 => [
+                    'id' => 245622414,
+                    'title' => 'Overhead view of a baby sleeping in the lap of his mother',
+                    'creator_id' => 224608,
+                    'creator_name' => 'Jacob Lund ',
+                    'creation_date' => '2019-01-26 23:42:40',
+                    'country_name' => 'Denmark',
+                    'thumbnail_url' => null,
+                    'thumbnail_html_tag' => null,
+                    'thumbnail_width' => null,
+                    'thumbnail_height' => null,
+                    'thumbnail_110_url' => null,
+                    'thumbnail_110_width' => null,
+                    'thumbnail_110_height' => null,
+                    'thumbnail_160_url' => null,
+                    'thumbnail_160_width' => null,
+                    'thumbnail_160_height' => null,
+                    'thumbnail_220_url' => null,
+                    'thumbnail_220_width' => null,
+                    'thumbnail_220_height' => null,
+                    'thumbnail_240_url' => 'http://image_url.jpg',
+                    'thumbnail_240_width' => null,
+                    'thumbnail_240_height' => null,
+                    'thumbnail_500_url' => 'http://image_url.jpg',
+                    'thumbnail_500_width' => null,
+                    'thumbnail_500_height' => null,
+                    'thumbnail_1000_url' => null,
+                    'thumbnail_1000_width' => null,
+                    'thumbnail_1000_height' => null,
+                    'media_type_id' => 1,
+                    'width' => 8008,
+                    'height' => 5339,
+                    'is_licensed' => null,
+                    'comp_url' => 'https://stock.adobe.com/Rest/Libraries/Watermarked/Download/245622414/2',
+                    'comp_width' => null,
+                    'comp_height' => null,
+                    'nb_views' => null,
+                    'nb_downloads' => null,
+                    'category' => [
+                        'id' => 695,
+                        'name' => 'People',
+                        'link' => null,
+                    ],
+                    'keywords' => [
+                        0 => [
+                            'name' => 'woman',
+                        ],
+                        1 => [
+                            'name' => 'boy',
+                        ],
+                        2 => [
+                            'name' => 'adult',
+                        ],
+                        3 => [
+                            'name' => 'children',
+                        ],
+                        4 => [
+                            'name' => 'baby',
+                        ],
+                        5 => [
+                            'name' => 'pair',
+                        ],
+                        6 => [
+                            'name' => 'mother',
+                        ],
+                        7 => [
+                            'name' => 'son',
+                        ],
+                        8 => [
+                            'name' => 'sleep',
+                        ],
+                        9 => [
+                            'name' => 'lay',
+                        ],
+                        10 => [
+                            'name' => 'to sit',
+                        ],
+                        11 => [
+                            'name' => 'lap',
+                        ],
+                        12 => [
+                            'name' => 'sofa',
+                        ],
+                        13 => [
+                            'name' => 'couch',
+                        ],
+                        14 => [
+                            'name' => 'casual attire',
+                        ],
+                        15 => [
+                            'name' => 'domestic',
+                        ],
+                        16 => [
+                            'name' => 'lifestyle',
+                        ],
+                        17 => [
+                            'name' => 'comfortable',
+                        ],
+                        18 => [
+                            'name' => 'safety',
+                        ],
+                        19 => [
+                            'name' => 'innocence',
+                        ],
+                        20 => [
+                            'name' => 'love',
+                        ],
+                        21 => [
+                            'name' => 'care',
+                        ],
+                        22 => [
+                            'name' => 'togetherness',
+                        ],
+                        23 => [
+                            'name' => 'indoor',
+                        ],
+                        24 => [
+                            'name' => 'day',
+                        ],
+                        25 => [
+                            'name' => '2',
+                        ],
+                        26 => [
+                            'name' => 'people',
+                        ],
+                    ],
+                    'has_releases' => null,
+                    'asset_type_id' => null,
+                    'vector_type' => null,
+                    'content_type' => 'image/jpeg',
+                    'framerate' => null,
+                    'duration' => null,
+                    'stock_id' => null,
+                    'comps' => null,
+                    'details_url' => 'http://image_url.jpg',
+                    'template_type_id' => null,
+                    'template_category_ids' => null,
+                    'marketing_text' => null,
+                    'description' => null,
+                    'size_bytes' => null,
+                    'premium_level_id' => 0,
+                    'is_premium' => null,
+                    'licenses' => null,
+                    'video_preview_url' => null,
+                    'VideoPreviewHeight' => null,
+                    'video_preview_width' => null,
+                    'video_preview_content_length' => null,
+                    'video_preview_content_type' => null,
+                    'video_small_preview_url' => null,
+                    'video_small_preview_height' => null,
+                    'video_small_preview_width' => null,
+                    'video_small_preview_content_length' => null,
+                    'video_small_preview_content_type' => null,
+                ],
+                2 => [
+                    'id' => 264327288,
+                    'title' => '子供　スケートボード　スケートパーク',
+                    'creator_id' => 206601882,
+                    'creator_name' => 'One',
+                    'creation_date' => '2019-04-25 06:25:16',
+                    'country_name' => 'Japan',
+                    'thumbnail_url' => null,
+                    'thumbnail_html_tag' => null,
+                    'thumbnail_width' => null,
+                    'thumbnail_height' => null,
+                    'thumbnail_110_url' => null,
+                    'thumbnail_110_width' => null,
+                    'thumbnail_110_height' => null,
+                    'thumbnail_160_url' => null,
+                    'thumbnail_160_width' => null,
+                    'thumbnail_160_height' => null,
+                    'thumbnail_220_url' => null,
+                    'thumbnail_220_width' => null,
+                    'thumbnail_220_height' => null,
+                    'thumbnail_240_url' => 'http://image_url.jpg',
+                    'thumbnail_240_width' => null,
+                    'thumbnail_240_height' => null,
+                    'thumbnail_500_url' => 'http://image_url.jpg',
+                    'thumbnail_500_width' => null,
+                    'thumbnail_500_height' => null,
+                    'thumbnail_1000_url' => null,
+                    'thumbnail_1000_width' => null,
+                    'thumbnail_1000_height' => null,
+                    'media_type_id' => 1,
+                    'width' => 8256,
+                    'height' => 5504,
+                    'is_licensed' => null,
+                    'comp_url' => 'https://stock.adobe.com/Rest/Libraries/Watermarked/Download/264327288/2',
+                    'comp_width' => null,
+                    'comp_height' => null,
+                    'nb_views' => null,
+                    'nb_downloads' => null,
+                    'category' => [
+                        'id' => 643,
+                        'name' => 'Lifestyle',
+                        'link' => null,
+                    ],
+                    'keywords' => [
+                        0 => [
+                            'name' => 'couple',
+                        ],
+                        1 => [
+                            'name' => 'woman',
+                        ],
+                        2 => [
+                            'name' => 'young',
+                        ],
+                        3 => [
+                            'name' => 'people',
+                        ],
+                        4 => [
+                            'name' => 'happy',
+                        ],
+                        5 => [
+                            'name' => 'car',
+                        ],
+                        6 => [
+                            'name' => 'family',
+                        ],
+                        7 => [
+                            'name' => 'smiling',
+                        ],
+                        8 => [
+                            'name' => 'mother',
+                        ],
+                        9 => [
+                            'name' => 'sitting',
+                        ],
+                        10 => [
+                            'name' => 'group',
+                        ],
+                        11 => [
+                            'name' => 'high school',
+                        ],
+                        12 => [
+                            'name' => 'shop',
+                        ],
+                        13 => [
+                            'name' => 'student',
+                        ],
+                        14 => [
+                            'name' => '2',
+                        ],
+                        15 => [
+                            'name' => 'business',
+                        ],
+                        16 => [
+                            'name' => 'together',
+                        ],
+                        17 => [
+                            'name' => 'love',
+                        ],
+                        18 => [
+                            'name' => 'smile',
+                        ],
+                        19 => [
+                            'name' => 'school',
+                        ],
+                        20 => [
+                            'name' => 'lovely',
+                        ],
+                        21 => [
+                            'name' => 'person',
+                        ],
+                        22 => [
+                            'name' => 'happiness',
+                        ],
+                        23 => [
+                            'name' => 'boy',
+                        ],
+                        24 => [
+                            'name' => 'lifestyle',
+                        ],
+                    ],
+                    'has_releases' => null,
+                    'asset_type_id' => null,
+                    'vector_type' => null,
+                    'content_type' => 'image/jpeg',
+                    'framerate' => null,
+                    'duration' => null,
+                    'stock_id' => null,
+                    'comps' => null,
+                    'details_url' => 'http://image_url.jpg',
+                    'template_type_id' => null,
+                    'template_category_ids' => null,
+                    'marketing_text' => null,
+                    'description' => null,
+                    'size_bytes' => null,
+                    'premium_level_id' => 0,
+                    'is_premium' => null,
+                    'licenses' => null,
+                    'video_preview_url' => null,
+                    'VideoPreviewHeight' => null,
+                    'video_preview_width' => null,
+                    'video_preview_content_length' => null,
+                    'video_preview_content_type' => null,
+                    'video_small_preview_url' => null,
+                    'video_small_preview_height' => null,
+                    'video_small_preview_width' => null,
+                    'video_small_preview_content_length' => null,
+                    'video_small_preview_content_type' => null,
+                ]
+            ],
+        ];
+    }
+}
