+++ AdobeStockClientStub/Model/Method/GetLicenseConfirmation.php
@@ -6,10 +6,10 @@ namespace Magento\AdobeStockClientStub\Model\Method;
 use AdobeStock\Api\Response\License;
 use Magento\AdobeStockClientApi\Api\Data\LicenseConfirmationInterface;
 use Magento\AdobeStockClientApi\Api\Data\LicenseConfirmationInterfaceFactory;
-use Magento\AdobeStockClientStub\Model\Template\LicenseConfirmation;
+use Magento\AdobeStockClientStub\Model\DataProvider\LicenseConfirmation;
 
 /**
- * Provides the stub data for the getLicenseConfirmation method of the AdobeStockClient
+ * Provides a stub data for the getLicenseConfirmation method of the AdobeStockClient
  */
 class GetLicenseConfirmation
 {
@@ -46,11 +46,9 @@ class GetLicenseConfirmation
      */
     public function execute(int $contentId): LicenseConfirmationInterface
     {
-        /** @var License $purchaseOptions */
-        $purchaseOptions = $this->licenseConfirmation->getLicenseStubObject($contentId);
+        $purchaseOptions = $this->licenseConfirmation->getLicenseStubObject($contentId)->getPurchaseOptions();
         $message = $purchaseOptions->getMessage();
         $canPurchase = $purchaseOptions->getPurchaseState() === 'possible';
-        /** @var LicenseConfirmationInterface $userQuota */
         $userQuota = $this->licenseConfirmationFactory->create();
         $userQuota->setMessage($message);
         $userQuota->setCanLicense($canPurchase);
@@ -58,4 +56,3 @@ class GetLicenseConfirmation
         return $userQuota;
     }
 }
-
