--- AdobeStockClientStub/Model/Method/SearchWithCriteria.php
@@ -1,24 +1,0 @@
-<?php
-/**
- * Copyright © Magento, Inc. All rights reserved.
- * See COPYING.txt for license details.
- */
-declare(strict_types=1);
-
-namespace Magento\AdobeStockClientStub\Model\Method;
-
-/**
- * Provides the stub result for the searchWithCriteria method of the AdobeStockClient
- */
-class SearchWithCriteria
-{
-    public function returnFirstPageSearchResult()
-    {
-        //@TODO add implementation
-    }
-
-    public function returnSecondPageSearchResult()
-    {
-        //@TODO add implementation
-    }
-}
