--- AdobeStockClientStub/Model/Template/RawStockResponse.php
@@ -1,4630 +1,0 @@
-<?php
-/**
- * Copyright © Magento, Inc. All rights reserved.
- * See COPYING.txt for license details.
- */
-declare(strict_types=1);
-
-namespace Magento\AdobeStockClientStub\Model\Template;
-
-/**
- * Represents the stub raw stock data response from the Adobe Stock service
- */
-class RawStockResponse
-{
-    /**
-     * Adobe Stock service raw response
-     *
-     * @return array
-     */
-    public function getRawAdobeStockResponse(): array
-    {
-        return [
-            'nb_results' => 427,
-            'files' => [
-                0 => [
-                    'id' => 244680548,
-                    'title' => 'Cheerful senior woman on a swing at a playground',
-                    'creator_id' => 204567087,
-                    'creator_name' => 'Rawpixel.com',
-                    'creation_date' => '2019-01-22 10:18:16',
-                    'country_name' => 'United Kingdom of Great Britain and Northern Ireland',
-                    'thumbnail_url' => null,
-                    'thumbnail_html_tag' => null,
-                    'thumbnail_width' => null,
-                    'thumbnail_height' => null,
-                    'thumbnail_110_url' => null,
-                    'thumbnail_110_width' => null,
-                    'thumbnail_110_height' => null,
-                    'thumbnail_160_url' => null,
-                    'thumbnail_160_width' => null,
-                    'thumbnail_160_height' => null,
-                    'thumbnail_220_url' => null,
-                    'thumbnail_220_width' => null,
-                    'thumbnail_220_height' => null,
-                    'thumbnail_240_url' => 'https://t4.ftcdn.net/jpg/02/44/68/05/240_F_244680548_toSihU4NKanDJqat4aqZU62gz5iShx3Q.jpg',
-                    'thumbnail_240_width' => null,
-                    'thumbnail_240_height' => null,
-                    'thumbnail_500_url' => 'https://as2.ftcdn.net/jpg/02/44/68/05/500_F_244680548_toSihU4NKanDJqat4aqZU62gz5iShx3Q.jpg',
-                    'thumbnail_500_width' => null,
-                    'thumbnail_500_height' => null,
-                    'thumbnail_1000_url' => null,
-                    'thumbnail_1000_width' => null,
-                    'thumbnail_1000_height' => null,
-                    'media_type_id' => 1,
-                    'width' => 5040,
-                    'height' => 3360,
-                    'is_licensed' => null,
-                    'comp_url' => 'https://stock.adobe.com/Rest/Libraries/Watermarked/Download/244680548/2',
-                    'comp_width' => null,
-                    'comp_height' => null,
-                    'nb_views' => null,
-                    'nb_downloads' => null,
-                    'category' => [
-                        'id' => 695,
-                        'name' => 'People',
-                        'link' => null,
-                    ],
-                    'keywords' => [
-                        0 => [
-                            'name' => 'woman',
-                        ],
-                        1 => [
-                            'name' => 'adult',
-                        ],
-                        2 => [
-                            'name' => 'senior',
-                        ],
-                        3 => [
-                            'name' => 'caucasian',
-                        ],
-                        4 => [
-                            'name' => 'alone',
-                        ],
-                        5 => [
-                            'name' => 'swing',
-                        ],
-                        6 => [
-                            'name' => 'playground',
-                        ],
-                        7 => [
-                            'name' => 'smile',
-                        ],
-                        8 => [
-                            'name' => 'white hair',
-                        ],
-                        9 => [
-                            'name' => 'shirt',
-                        ],
-                        10 => [
-                            'name' => 'chain',
-                        ],
-                        11 => [
-                            'name' => 'park',
-                        ],
-                        12 => [
-                            'name' => 'fun',
-                        ],
-                        13 => [
-                            'name' => 'cheerful',
-                        ],
-                        14 => [
-                            'name' => 'vitality',
-                        ],
-                        15 => [
-                            'name' => 'retired',
-                        ],
-                        16 => [
-                            'name' => 'happiness',
-                        ],
-                        17 => [
-                            'name' => 'enjoyment',
-                        ],
-                        18 => [
-                            'name' => 'outdoors',
-                        ],
-                        19 => [
-                            'name' => 'day',
-                        ],
-                        20 => [
-                            'name' => '1',
-                        ],
-                        21 => [
-                            'name' => 'people',
-                        ],
-                    ],
-                    'has_releases' => null,
-                    'asset_type_id' => null,
-                    'vector_type' => null,
-                    'content_type' => 'image/jpeg',
-                    'framerate' => null,
-                    'duration' => null,
-                    'stock_id' => null,
-                    'comps' => null,
-                    'details_url' => 'https://stock.adobe.com/244680548?as_channel=affiliate&as_source=api&as_content=7f387d439e4f4e64ae42060fc571c456',
-                    'template_type_id' => null,
-                    'template_category_ids' => null,
-                    'marketing_text' => null,
-                    'description' => null,
-                    'size_bytes' => null,
-                    'premium_level_id' => 0,
-                    'is_premium' => null,
-                    'licenses' => null,
-                    'video_preview_url' => null,
-                    'VideoPreviewHeight' => null,
-                    'video_preview_width' => null,
-                    'video_preview_content_length' => null,
-                    'video_preview_content_type' => null,
-                    'video_small_preview_url' => null,
-                    'video_small_preview_height' => null,
-                    'video_small_preview_width' => null,
-                    'video_small_preview_content_length' => null,
-                    'video_small_preview_content_type' => null,
-                ],
-                1 => [
-                    'id' => 245622414,
-                    'title' => 'Overhead view of a baby sleeping in the lap of his mother',
-                    'creator_id' => 224608,
-                    'creator_name' => 'Jacob Lund ',
-                    'creation_date' => '2019-01-26 23:42:40',
-                    'country_name' => 'Denmark',
-                    'thumbnail_url' => null,
-                    'thumbnail_html_tag' => null,
-                    'thumbnail_width' => null,
-                    'thumbnail_height' => null,
-                    'thumbnail_110_url' => null,
-                    'thumbnail_110_width' => null,
-                    'thumbnail_110_height' => null,
-                    'thumbnail_160_url' => null,
-                    'thumbnail_160_width' => null,
-                    'thumbnail_160_height' => null,
-                    'thumbnail_220_url' => null,
-                    'thumbnail_220_width' => null,
-                    'thumbnail_220_height' => null,
-                    'thumbnail_240_url' => 'https://t3.ftcdn.net/jpg/02/45/62/24/240_F_245622414_ZCqonGBDteNu34rZ2VIT0KXHdbg0RKVr.jpg',
-                    'thumbnail_240_width' => null,
-                    'thumbnail_240_height' => null,
-                    'thumbnail_500_url' => 'https://as1.ftcdn.net/jpg/02/45/62/24/500_F_245622414_ZCqonGBDteNu34rZ2VIT0KXHdbg0RKVr.jpg',
-                    'thumbnail_500_width' => null,
-                    'thumbnail_500_height' => null,
-                    'thumbnail_1000_url' => null,
-                    'thumbnail_1000_width' => null,
-                    'thumbnail_1000_height' => null,
-                    'media_type_id' => 1,
-                    'width' => 8008,
-                    'height' => 5339,
-                    'is_licensed' => null,
-                    'comp_url' => 'https://stock.adobe.com/Rest/Libraries/Watermarked/Download/245622414/2',
-                    'comp_width' => null,
-                    'comp_height' => null,
-                    'nb_views' => null,
-                    'nb_downloads' => null,
-                    'category' => [
-                        'id' => 695,
-                        'name' => 'People',
-                        'link' => null,
-                    ],
-                    'keywords' => [
-                        0 => [
-                            'name' => 'woman',
-                        ],
-                        1 => [
-                            'name' => 'boy',
-                        ],
-                        2 => [
-                            'name' => 'adult',
-                        ],
-                        3 => [
-                            'name' => 'children',
-                        ],
-                        4 => [
-                            'name' => 'baby',
-                        ],
-                        5 => [
-                            'name' => 'pair',
-                        ],
-                        6 => [
-                            'name' => 'mother',
-                        ],
-                        7 => [
-                            'name' => 'son',
-                        ],
-                        8 => [
-                            'name' => 'sleep',
-                        ],
-                        9 => [
-                            'name' => 'lay',
-                        ],
-                        10 => [
-                            'name' => 'to sit',
-                        ],
-                        11 => [
-                            'name' => 'lap',
-                        ],
-                        12 => [
-                            'name' => 'sofa',
-                        ],
-                        13 => [
-                            'name' => 'couch',
-                        ],
-                        14 => [
-                            'name' => 'casual attire',
-                        ],
-                        15 => [
-                            'name' => 'domestic',
-                        ],
-                        16 => [
-                            'name' => 'lifestyle',
-                        ],
-                        17 => [
-                            'name' => 'comfortable',
-                        ],
-                        18 => [
-                            'name' => 'safety',
-                        ],
-                        19 => [
-                            'name' => 'innocence',
-                        ],
-                        20 => [
-                            'name' => 'love',
-                        ],
-                        21 => [
-                            'name' => 'care',
-                        ],
-                        22 => [
-                            'name' => 'togetherness',
-                        ],
-                        23 => [
-                            'name' => 'indoor',
-                        ],
-                        24 => [
-                            'name' => 'day',
-                        ],
-                        25 => [
-                            'name' => '2',
-                        ],
-                        26 => [
-                            'name' => 'people',
-                        ],
-                    ],
-                    'has_releases' => null,
-                    'asset_type_id' => null,
-                    'vector_type' => null,
-                    'content_type' => 'image/jpeg',
-                    'framerate' => null,
-                    'duration' => null,
-                    'stock_id' => null,
-                    'comps' => null,
-                    'details_url' => 'https://stock.adobe.com/245622414?as_channel=affiliate&as_source=api&as_content=7f387d439e4f4e64ae42060fc571c456',
-                    'template_type_id' => null,
-                    'template_category_ids' => null,
-                    'marketing_text' => null,
-                    'description' => null,
-                    'size_bytes' => null,
-                    'premium_level_id' => 0,
-                    'is_premium' => null,
-                    'licenses' => null,
-                    'video_preview_url' => null,
-                    'VideoPreviewHeight' => null,
-                    'video_preview_width' => null,
-                    'video_preview_content_length' => null,
-                    'video_preview_content_type' => null,
-                    'video_small_preview_url' => null,
-                    'video_small_preview_height' => null,
-                    'video_small_preview_width' => null,
-                    'video_small_preview_content_length' => null,
-                    'video_small_preview_content_type' => null,
-                ],
-                2 => [
-                    'id' => 264327288,
-                    'title' => '子供　スケートボード　スケートパーク',
-                    'creator_id' => 206601882,
-                    'creator_name' => 'One',
-                    'creation_date' => '2019-04-25 06:25:16',
-                    'country_name' => 'Japan',
-                    'thumbnail_url' => null,
-                    'thumbnail_html_tag' => null,
-                    'thumbnail_width' => null,
-                    'thumbnail_height' => null,
-                    'thumbnail_110_url' => null,
-                    'thumbnail_110_width' => null,
-                    'thumbnail_110_height' => null,
-                    'thumbnail_160_url' => null,
-                    'thumbnail_160_width' => null,
-                    'thumbnail_160_height' => null,
-                    'thumbnail_220_url' => null,
-                    'thumbnail_220_width' => null,
-                    'thumbnail_220_height' => null,
-                    'thumbnail_240_url' => 'https://t3.ftcdn.net/jpg/02/64/32/72/240_F_264327288_xzYWfhL6QsIjs5Y60EKtOcjTOZQcLLo8.jpg',
-                    'thumbnail_240_width' => null,
-                    'thumbnail_240_height' => null,
-                    'thumbnail_500_url' => 'https://as1.ftcdn.net/jpg/02/64/32/72/500_F_264327288_xzYWfhL6QsIjs5Y60EKtOcjTOZQcLLo8.jpg',
-                    'thumbnail_500_width' => null,
-                    'thumbnail_500_height' => null,
-                    'thumbnail_1000_url' => null,
-                    'thumbnail_1000_width' => null,
-                    'thumbnail_1000_height' => null,
-                    'media_type_id' => 1,
-                    'width' => 8256,
-                    'height' => 5504,
-                    'is_licensed' => null,
-                    'comp_url' => 'https://stock.adobe.com/Rest/Libraries/Watermarked/Download/264327288/2',
-                    'comp_width' => null,
-                    'comp_height' => null,
-                    'nb_views' => null,
-                    'nb_downloads' => null,
-                    'category' => [
-                        'id' => 643,
-                        'name' => 'Lifestyle',
-                        'link' => null,
-                    ],
-                    'keywords' => [
-                        0 => [
-                            'name' => 'couple',
-                        ],
-                        1 => [
-                            'name' => 'woman',
-                        ],
-                        2 => [
-                            'name' => 'young',
-                        ],
-                        3 => [
-                            'name' => 'people',
-                        ],
-                        4 => [
-                            'name' => 'happy',
-                        ],
-                        5 => [
-                            'name' => 'car',
-                        ],
-                        6 => [
-                            'name' => 'family',
-                        ],
-                        7 => [
-                            'name' => 'smiling',
-                        ],
-                        8 => [
-                            'name' => 'mother',
-                        ],
-                        9 => [
-                            'name' => 'sitting',
-                        ],
-                        10 => [
-                            'name' => 'group',
-                        ],
-                        11 => [
-                            'name' => 'high school',
-                        ],
-                        12 => [
-                            'name' => 'shop',
-                        ],
-                        13 => [
-                            'name' => 'student',
-                        ],
-                        14 => [
-                            'name' => '2',
-                        ],
-                        15 => [
-                            'name' => 'business',
-                        ],
-                        16 => [
-                            'name' => 'together',
-                        ],
-                        17 => [
-                            'name' => 'love',
-                        ],
-                        18 => [
-                            'name' => 'smile',
-                        ],
-                        19 => [
-                            'name' => 'school',
-                        ],
-                        20 => [
-                            'name' => 'lovely',
-                        ],
-                        21 => [
-                            'name' => 'person',
-                        ],
-                        22 => [
-                            'name' => 'happiness',
-                        ],
-                        23 => [
-                            'name' => 'boy',
-                        ],
-                        24 => [
-                            'name' => 'lifestyle',
-                        ],
-                    ],
-                    'has_releases' => null,
-                    'asset_type_id' => null,
-                    'vector_type' => null,
-                    'content_type' => 'image/jpeg',
-                    'framerate' => null,
-                    'duration' => null,
-                    'stock_id' => null,
-                    'comps' => null,
-                    'details_url' => 'https://stock.adobe.com/264327288?as_channel=affiliate&as_source=api&as_content=7f387d439e4f4e64ae42060fc571c456',
-                    'template_type_id' => null,
-                    'template_category_ids' => null,
-                    'marketing_text' => null,
-                    'description' => null,
-                    'size_bytes' => null,
-                    'premium_level_id' => 0,
-                    'is_premium' => null,
-                    'licenses' => null,
-                    'video_preview_url' => null,
-                    'VideoPreviewHeight' => null,
-                    'video_preview_width' => null,
-                    'video_preview_content_length' => null,
-                    'video_preview_content_type' => null,
-                    'video_small_preview_url' => null,
-                    'video_small_preview_height' => null,
-                    'video_small_preview_width' => null,
-                    'video_small_preview_content_length' => null,
-                    'video_small_preview_content_type' => null,
-                ],
-                3 => [
-                    'id' => 264874647,
-                    'title' => 'Overhead view of highway interchange',
-                    'creator_id' => 290675,
-                    'creator_name' => 'Felix Mizioznikov',
-                    'creation_date' => '2019-04-28 03:44:53',
-                    'country_name' => 'United States of America',
-                    'thumbnail_url' => null,
-                    'thumbnail_html_tag' => null,
-                    'thumbnail_width' => null,
-                    'thumbnail_height' => null,
-                    'thumbnail_110_url' => null,
-                    'thumbnail_110_width' => null,
-                    'thumbnail_110_height' => null,
-                    'thumbnail_160_url' => null,
-                    'thumbnail_160_width' => null,
-                    'thumbnail_160_height' => null,
-                    'thumbnail_220_url' => null,
-                    'thumbnail_220_width' => null,
-                    'thumbnail_220_height' => null,
-                    'thumbnail_240_url' => 'https://t3.ftcdn.net/jpg/02/64/87/46/240_F_264874647_AlOi4yQBYXzsYFUZ1uaXcY0doGIWkADl.jpg',
-                    'thumbnail_240_width' => null,
-                    'thumbnail_240_height' => null,
-                    'thumbnail_500_url' => 'https://as1.ftcdn.net/jpg/02/64/87/46/500_F_264874647_AlOi4yQBYXzsYFUZ1uaXcY0doGIWkADl.jpg',
-                    'thumbnail_500_width' => null,
-                    'thumbnail_500_height' => null,
-                    'thumbnail_1000_url' => null,
-                    'thumbnail_1000_width' => null,
-                    'thumbnail_1000_height' => null,
-                    'media_type_id' => 1,
-                    'width' => 5464,
-                    'height' => 3640,
-                    'is_licensed' => null,
-                    'comp_url' => 'https://stock.adobe.com/Rest/Libraries/Watermarked/Download/264874647/2',
-                    'comp_width' => null,
-                    'comp_height' => null,
-                    'nb_views' => null,
-                    'nb_downloads' => null,
-                    'category' => [
-                        'id' => 96,
-                        'name' => 'Buildings and Architecture',
-                        'link' => null,
-                    ],
-                    'keywords' => [
-                        0 => [
-                            'name' => 'highway',
-                        ],
-                        1 => [
-                            'name' => 'expressway',
-                        ],
-                        2 => [
-                            'name' => 'interchange',
-                        ],
-                        3 => [
-                            'name' => 'road',
-                        ],
-                        4 => [
-                            'name' => 'car',
-                        ],
-                        5 => [
-                            'name' => 'auto',
-                        ],
-                        6 => [
-                            'name' => 'automobile',
-                        ],
-                        7 => [
-                            'name' => 'overpass',
-                        ],
-                        8 => [
-                            'name' => 'direction',
-                        ],
-                        9 => [
-                            'name' => 'transportation',
-                        ],
-                        10 => [
-                            'name' => 'travel',
-                        ],
-                        11 => [
-                            'name' => 'complexity',
-                        ],
-                        12 => [
-                            'name' => 'curve',
-                        ],
-                        13 => [
-                            'name' => 'aerial',
-                        ],
-                        14 => [
-                            'name' => 'outdoors',
-                        ],
-                        15 => [
-                            'name' => 'day',
-                        ],
-                        16 => [
-                            'name' => 'no people',
-                        ],
-                        17 => [
-                            'name' => 'nobody',
-                        ],
-                    ],
-                    'has_releases' => null,
-                    'asset_type_id' => null,
-                    'vector_type' => null,
-                    'content_type' => 'image/jpeg',
-                    'framerate' => null,
-                    'duration' => null,
-                    'stock_id' => null,
-                    'comps' => null,
-                    'details_url' => 'https://stock.adobe.com/264874647?as_channel=affiliate&as_source=api&as_content=7f387d439e4f4e64ae42060fc571c456',
-                    'template_type_id' => null,
-                    'template_category_ids' => null,
-                    'marketing_text' => null,
-                    'description' => null,
-                    'size_bytes' => null,
-                    'premium_level_id' => 0,
-                    'is_premium' => null,
-                    'licenses' => null,
-                    'video_preview_url' => null,
-                    'VideoPreviewHeight' => null,
-                    'video_preview_width' => null,
-                    'video_preview_content_length' => null,
-                    'video_preview_content_type' => null,
-                    'video_small_preview_url' => null,
-                    'video_small_preview_height' => null,
-                    'video_small_preview_width' => null,
-                    'video_small_preview_content_length' => null,
-                    'video_small_preview_content_type' => null,
-                ],
-                4 => [
-                    'id' => 295213410,
-                    'title' => 'Overhead view of two women lying on pebbles, covering eyes',
-                    'creator_id' => 201027626,
-                    'creator_name' => 'Smile',
-                    'creation_date' => '2019-10-11 05:01:41',
-                    'country_name' => 'Russian Federation',
-                    'thumbnail_url' => null,
-                    'thumbnail_html_tag' => null,
-                    'thumbnail_width' => null,
-                    'thumbnail_height' => null,
-                    'thumbnail_110_url' => null,
-                    'thumbnail_110_width' => null,
-                    'thumbnail_110_height' => null,
-                    'thumbnail_160_url' => null,
-                    'thumbnail_160_width' => null,
-                    'thumbnail_160_height' => null,
-                    'thumbnail_220_url' => null,
-                    'thumbnail_220_width' => null,
-                    'thumbnail_220_height' => null,
-                    'thumbnail_240_url' => 'https://t3.ftcdn.net/jpg/02/95/21/34/240_F_295213410_8miDWqkRWnO3V9jUV5Kpi8CiRAeg3jR2.jpg',
-                    'thumbnail_240_width' => null,
-                    'thumbnail_240_height' => null,
-                    'thumbnail_500_url' => 'https://as1.ftcdn.net/jpg/02/95/21/34/500_F_295213410_8miDWqkRWnO3V9jUV5Kpi8CiRAeg3jR2.jpg',
-                    'thumbnail_500_width' => null,
-                    'thumbnail_500_height' => null,
-                    'thumbnail_1000_url' => null,
-                    'thumbnail_1000_width' => null,
-                    'thumbnail_1000_height' => null,
-                    'media_type_id' => 1,
-                    'width' => 3840,
-                    'height' => 5760,
-                    'is_licensed' => null,
-                    'comp_url' => 'https://stock.adobe.com/Rest/Libraries/Watermarked/Download/295213410/2',
-                    'comp_width' => null,
-                    'comp_height' => null,
-                    'nb_views' => null,
-                    'nb_downloads' => null,
-                    'category' => [
-                        'id' => 695,
-                        'name' => 'People',
-                        'link' => null,
-                    ],
-                    'keywords' => [
-                        0 => [
-                            'name' => 'woman',
-                        ],
-                        1 => [
-                            'name' => 'adult',
-                        ],
-                        2 => [
-                            'name' => 'young',
-                        ],
-                        3 => [
-                            'name' => 'pair',
-                        ],
-                        4 => [
-                            'name' => 'lay',
-                        ],
-                        5 => [
-                            'name' => 'pebble',
-                        ],
-                        6 => [
-                            'name' => 'blond',
-                        ],
-                        7 => [
-                            'name' => 'brunette',
-                        ],
-                        8 => [
-                            'name' => 'long hair',
-                        ],
-                        9 => [
-                            'name' => 'cover',
-                        ],
-                        10 => [
-                            'name' => 'hand',
-                        ],
-                        11 => [
-                            'name' => 'eye',
-                        ],
-                        12 => [
-                            'name' => 'dress',
-                        ],
-                        13 => [
-                            'name' => 'pattern',
-                        ],
-                        14 => [
-                            'name' => 'closeness',
-                        ],
-                        15 => [
-                            'name' => 'togetherness',
-                        ],
-                        16 => [
-                            'name' => 'friendship',
-                        ],
-                        17 => [
-                            'name' => 'share',
-                        ],
-                        18 => [
-                            'name' => 'outdoors',
-                        ],
-                        19 => [
-                            'name' => 'day',
-                        ],
-                        20 => [
-                            'name' => '2',
-                        ],
-                        21 => [
-                            'name' => 'people',
-                        ],
-                    ],
-                    'has_releases' => null,
-                    'asset_type_id' => null,
-                    'vector_type' => null,
-                    'content_type' => 'image/jpeg',
-                    'framerate' => null,
-                    'duration' => null,
-                    'stock_id' => null,
-                    'comps' => null,
-                    'details_url' => 'https://stock.adobe.com/295213410?as_channel=affiliate&as_source=api&as_content=7f387d439e4f4e64ae42060fc571c456',
-                    'template_type_id' => null,
-                    'template_category_ids' => null,
-                    'marketing_text' => null,
-                    'description' => null,
-                    'size_bytes' => null,
-                    'premium_level_id' => 0,
-                    'is_premium' => null,
-                    'licenses' => null,
-                    'video_preview_url' => null,
-                    'VideoPreviewHeight' => null,
-                    'video_preview_width' => null,
-                    'video_preview_content_length' => null,
-                    'video_preview_content_type' => null,
-                    'video_small_preview_url' => null,
-                    'video_small_preview_height' => null,
-                    'video_small_preview_width' => null,
-                    'video_small_preview_content_length' => null,
-                    'video_small_preview_content_type' => null,
-                ],
-                5 => [
-                    'id' => 276132997,
-                    'title' => 'Profiles of two female surfers on beach in sunlight',
-                    'creator_id' => 224608,
-                    'creator_name' => 'Jacob Lund ',
-                    'creation_date' => '2019-07-01 14:20:16',
-                    'country_name' => 'Denmark',
-                    'thumbnail_url' => null,
-                    'thumbnail_html_tag' => null,
-                    'thumbnail_width' => null,
-                    'thumbnail_height' => null,
-                    'thumbnail_110_url' => null,
-                    'thumbnail_110_width' => null,
-                    'thumbnail_110_height' => null,
-                    'thumbnail_160_url' => null,
-                    'thumbnail_160_width' => null,
-                    'thumbnail_160_height' => null,
-                    'thumbnail_220_url' => null,
-                    'thumbnail_220_width' => null,
-                    'thumbnail_220_height' => null,
-                    'thumbnail_240_url' => 'https://t4.ftcdn.net/jpg/02/76/13/29/240_F_276132997_xAxX96gXANvGmRilrtu4OVXONclBgTwT.jpg',
-                    'thumbnail_240_width' => null,
-                    'thumbnail_240_height' => null,
-                    'thumbnail_500_url' => 'https://as2.ftcdn.net/jpg/02/76/13/29/500_F_276132997_xAxX96gXANvGmRilrtu4OVXONclBgTwT.jpg',
-                    'thumbnail_500_width' => null,
-                    'thumbnail_500_height' => null,
-                    'thumbnail_1000_url' => null,
-                    'thumbnail_1000_width' => null,
-                    'thumbnail_1000_height' => null,
-                    'media_type_id' => 1,
-                    'width' => 8688,
-                    'height' => 5792,
-                    'is_licensed' => null,
-                    'comp_url' => 'https://stock.adobe.com/Rest/Libraries/Watermarked/Download/276132997/2',
-                    'comp_width' => null,
-                    'comp_height' => null,
-                    'nb_views' => null,
-                    'nb_downloads' => null,
-                    'category' => [
-                        'id' => 695,
-                        'name' => 'People',
-                        'link' => null,
-                    ],
-                    'keywords' => [
-                        0 => [
-                            'name' => 'woman',
-                        ],
-                        1 => [
-                            'name' => 'adult',
-                        ],
-                        2 => [
-                            'name' => 'young',
-                        ],
-                        3 => [
-                            'name' => 'pair',
-                        ],
-                        4 => [
-                            'name' => 'surfer',
-                        ],
-                        5 => [
-                            'name' => 'surfboard',
-                        ],
-                        6 => [
-                            'name' => 'beach',
-                        ],
-                        7 => [
-                            'name' => 'sunlight',
-                        ],
-                        8 => [
-                            'name' => 'caucasian',
-                        ],
-                        9 => [
-                            'name' => 'surfing',
-                        ],
-                        10 => [
-                            'name' => 'blond',
-                        ],
-                        11 => [
-                            'name' => 'long hair',
-                        ],
-                        12 => [
-                            'name' => 'summer',
-                        ],
-                        13 => [
-                            'name' => 'vacation',
-                        ],
-                        14 => [
-                            'name' => 'profile',
-                        ],
-                        15 => [
-                            'name' => 'relaxation',
-                        ],
-                        16 => [
-                            'name' => 'lifestyle',
-                        ],
-                        17 => [
-                            'name' => 'selective focus',
-                        ],
-                        18 => [
-                            'name' => 'outdoors',
-                        ],
-                        19 => [
-                            'name' => 'day',
-                        ],
-                        20 => [
-                            'name' => '2',
-                        ],
-                        21 => [
-                            'name' => 'people',
-                        ],
-                    ],
-                    'has_releases' => null,
-                    'asset_type_id' => null,
-                    'vector_type' => null,
-                    'content_type' => 'image/jpeg',
-                    'framerate' => null,
-                    'duration' => null,
-                    'stock_id' => null,
-                    'comps' => null,
-                    'details_url' => 'https://stock.adobe.com/276132997?as_channel=affiliate&as_source=api&as_content=7f387d439e4f4e64ae42060fc571c456',
-                    'template_type_id' => null,
-                    'template_category_ids' => null,
-                    'marketing_text' => null,
-                    'description' => null,
-                    'size_bytes' => null,
-                    'premium_level_id' => 0,
-                    'is_premium' => null,
-                    'licenses' => null,
-                    'video_preview_url' => null,
-                    'VideoPreviewHeight' => null,
-                    'video_preview_width' => null,
-                    'video_preview_content_length' => null,
-                    'video_preview_content_type' => null,
-                    'video_small_preview_url' => null,
-                    'video_small_preview_height' => null,
-                    'video_small_preview_width' => null,
-                    'video_small_preview_content_length' => null,
-                    'video_small_preview_content_type' => null,
-                ],
-                6 => [
-                    'id' => 285504582,
-                    'title' => 'Father and son flying kite in field',
-                    'creator_id' => 204337425,
-                    'creator_name' => 'rh2010',
-                    'creation_date' => '2019-08-22 18:55:06',
-                    'country_name' => 'Ukraine',
-                    'thumbnail_url' => null,
-                    'thumbnail_html_tag' => null,
-                    'thumbnail_width' => null,
-                    'thumbnail_height' => null,
-                    'thumbnail_110_url' => null,
-                    'thumbnail_110_width' => null,
-                    'thumbnail_110_height' => null,
-                    'thumbnail_160_url' => null,
-                    'thumbnail_160_width' => null,
-                    'thumbnail_160_height' => null,
-                    'thumbnail_220_url' => null,
-                    'thumbnail_220_width' => null,
-                    'thumbnail_220_height' => null,
-                    'thumbnail_240_url' => 'https://t4.ftcdn.net/jpg/02/85/50/45/240_F_285504582_gAJq9MBr5FjSbXoAJTdR6emv5bLmEHdk.jpg',
-                    'thumbnail_240_width' => null,
-                    'thumbnail_240_height' => null,
-                    'thumbnail_500_url' => 'https://as2.ftcdn.net/jpg/02/85/50/45/500_F_285504582_gAJq9MBr5FjSbXoAJTdR6emv5bLmEHdk.jpg',
-                    'thumbnail_500_width' => null,
-                    'thumbnail_500_height' => null,
-                    'thumbnail_1000_url' => null,
-                    'thumbnail_1000_width' => null,
-                    'thumbnail_1000_height' => null,
-                    'media_type_id' => 1,
-                    'width' => 7360,
-                    'height' => 4912,
-                    'is_licensed' => null,
-                    'comp_url' => 'https://stock.adobe.com/Rest/Libraries/Watermarked/Download/285504582/2',
-                    'comp_width' => null,
-                    'comp_height' => null,
-                    'nb_views' => null,
-                    'nb_downloads' => null,
-                    'category' => [
-                        'id' => 498,
-                        'name' => 'Hobbies and Leisure',
-                        'link' => null,
-                    ],
-                    'keywords' => [
-                        0 => [
-                            'name' => 'man',
-                        ],
-                        1 => [
-                            'name' => 'boy',
-                        ],
-                        2 => [
-                            'name' => 'adult',
-                        ],
-                        3 => [
-                            'name' => 'children',
-                        ],
-                        4 => [
-                            'name' => 'young',
-                        ],
-                        5 => [
-                            'name' => 'elementary',
-                        ],
-                        6 => [
-                            'name' => 'caucasian',
-                        ],
-                        7 => [
-                            'name' => 'pair',
-                        ],
-                        8 => [
-                            'name' => 'father',
-                        ],
-                        9 => [
-                            'name' => 'son',
-                        ],
-                        10 => [
-                            'name' => 'kite',
-                        ],
-                        11 => [
-                            'name' => 'fly',
-                        ],
-                        12 => [
-                            'name' => 'field',
-                        ],
-                        13 => [
-                            'name' => 'stubble field',
-                        ],
-                        14 => [
-                            'name' => 'run',
-                        ],
-                        15 => [
-                            'name' => 'play',
-                        ],
-                        16 => [
-                            'name' => 'casual attire',
-                        ],
-                        17 => [
-                            'name' => 'hat',
-                        ],
-                        18 => [
-                            'name' => 'sky',
-                        ],
-                        19 => [
-                            'name' => 'many-coloured',
-                        ],
-                        20 => [
-                            'name' => 'colourful',
-                        ],
-                        21 => [
-                            'name' => 'childhood',
-                        ],
-                        22 => [
-                            'name' => 'enjoyment',
-                        ],
-                        23 => [
-                            'name' => 'togetherness',
-                        ],
-                        24 => [
-                            'name' => 'carefree',
-                        ],
-                        25 => [
-                            'name' => 'vacation',
-                        ],
-                        26 => [
-                            'name' => 'happiness',
-                        ],
-                        27 => [
-                            'name' => 'lifestyle',
-                        ],
-                        28 => [
-                            'name' => 'outdoors',
-                        ],
-                        29 => [
-                            'name' => 'day',
-                        ],
-                        30 => [
-                            'name' => '2',
-                        ],
-                        31 => [
-                            'name' => 'people',
-                        ],
-                    ],
-                    'has_releases' => null,
-                    'asset_type_id' => null,
-                    'vector_type' => null,
-                    'content_type' => 'image/jpeg',
-                    'framerate' => null,
-                    'duration' => null,
-                    'stock_id' => null,
-                    'comps' => null,
-                    'details_url' => 'https://stock.adobe.com/285504582?as_channel=affiliate&as_source=api&as_content=7f387d439e4f4e64ae42060fc571c456',
-                    'template_type_id' => null,
-                    'template_category_ids' => null,
-                    'marketing_text' => null,
-                    'description' => null,
-                    'size_bytes' => null,
-                    'premium_level_id' => 0,
-                    'is_premium' => null,
-                    'licenses' => null,
-                    'video_preview_url' => null,
-                    'VideoPreviewHeight' => null,
-                    'video_preview_width' => null,
-                    'video_preview_content_length' => null,
-                    'video_preview_content_type' => null,
-                    'video_small_preview_url' => null,
-                    'video_small_preview_height' => null,
-                    'video_small_preview_width' => null,
-                    'video_small_preview_content_length' => null,
-                    'video_small_preview_content_type' => null,
-                ],
-                7 => [
-                    'id' => 271858149,
-                    'title' => 'Soldier in wheelchair with American flag on airport',
-                    'creator_id' => 205191830,
-                    'creator_name' => 'Yakobchuk Olena',
-                    'creation_date' => '2019-06-06 23:51:21',
-                    'country_name' => 'Ukraine',
-                    'thumbnail_url' => null,
-                    'thumbnail_html_tag' => null,
-                    'thumbnail_width' => null,
-                    'thumbnail_height' => null,
-                    'thumbnail_110_url' => null,
-                    'thumbnail_110_width' => null,
-                    'thumbnail_110_height' => null,
-                    'thumbnail_160_url' => null,
-                    'thumbnail_160_width' => null,
-                    'thumbnail_160_height' => null,
-                    'thumbnail_220_url' => null,
-                    'thumbnail_220_width' => null,
-                    'thumbnail_220_height' => null,
-                    'thumbnail_240_url' => 'https://t4.ftcdn.net/jpg/02/71/85/81/240_F_271858149_XYVB83JchPEgPYl5IoZDu94allYgUXdr.jpg',
-                    'thumbnail_240_width' => null,
-                    'thumbnail_240_height' => null,
-                    'thumbnail_500_url' => 'https://as2.ftcdn.net/jpg/02/71/85/81/500_F_271858149_XYVB83JchPEgPYl5IoZDu94allYgUXdr.jpg',
-                    'thumbnail_500_width' => null,
-                    'thumbnail_500_height' => null,
-                    'thumbnail_1000_url' => null,
-                    'thumbnail_1000_width' => null,
-                    'thumbnail_1000_height' => null,
-                    'media_type_id' => 1,
-                    'width' => 5533,
-                    'height' => 3694,
-                    'is_licensed' => null,
-                    'comp_url' => 'https://stock.adobe.com/Rest/Libraries/Watermarked/Download/271858149/2',
-                    'comp_width' => null,
-                    'comp_height' => null,
-                    'nb_views' => null,
-                    'nb_downloads' => null,
-                    'category' => [
-                        'id' => 695,
-                        'name' => 'People',
-                        'link' => null,
-                    ],
-                    'keywords' => [
-                        0 => [
-                            'name' => 'man',
-                        ],
-                        1 => [
-                            'name' => 'adult',
-                        ],
-                        2 => [
-                            'name' => 'mid',
-                        ],
-                        3 => [
-                            'name' => 'caucasian',
-                        ],
-                        4 => [
-                            'name' => 'alone',
-                        ],
-                        5 => [
-                            'name' => 'soldier',
-                        ],
-                        6 => [
-                            'name' => 'wheelchair',
-                        ],
-                        7 => [
-                            'name' => 'american',
-                        ],
-                        8 => [
-                            'name' => 'flag',
-                        ],
-                        9 => [
-                            'name' => 'aerodrome',
-                        ],
-                        10 => [
-                            'name' => 'army',
-                        ],
-                        11 => [
-                            'name' => 'military',
-                        ],
-                        12 => [
-                            'name' => 'patriotism',
-                        ],
-                        13 => [
-                            'name' => 'homecoming',
-                        ],
-                        14 => [
-                            'name' => 'identity',
-                        ],
-                        15 => [
-                            'name' => 'disabled',
-                        ],
-                        16 => [
-                            'name' => 'disability',
-                        ],
-                        17 => [
-                            'name' => 'dedication',
-                        ],
-                        18 => [
-                            'name' => 'travel',
-                        ],
-                        19 => [
-                            'name' => 'solitude',
-                        ],
-                        20 => [
-                            'name' => 'lifestyle',
-                        ],
-                        21 => [
-                            'name' => 'window',
-                        ],
-                        22 => [
-                            'name' => 'cloud',
-                        ],
-                        23 => [
-                            'name' => 'to sit',
-                        ],
-                        24 => [
-                            'name' => 'indoor',
-                        ],
-                        25 => [
-                            'name' => 'day',
-                        ],
-                        26 => [
-                            'name' => '1',
-                        ],
-                        27 => [
-                            'name' => 'people',
-                        ],
-                    ],
-                    'has_releases' => null,
-                    'asset_type_id' => null,
-                    'vector_type' => null,
-                    'content_type' => 'image/jpeg',
-                    'framerate' => null,
-                    'duration' => null,
-                    'stock_id' => null,
-                    'comps' => null,
-                    'details_url' => 'https://stock.adobe.com/271858149?as_channel=affiliate&as_source=api&as_content=7f387d439e4f4e64ae42060fc571c456',
-                    'template_type_id' => null,
-                    'template_category_ids' => null,
-                    'marketing_text' => null,
-                    'description' => null,
-                    'size_bytes' => null,
-                    'premium_level_id' => 0,
-                    'is_premium' => null,
-                    'licenses' => null,
-                    'video_preview_url' => null,
-                    'VideoPreviewHeight' => null,
-                    'video_preview_width' => null,
-                    'video_preview_content_length' => null,
-                    'video_preview_content_type' => null,
-                    'video_small_preview_url' => null,
-                    'video_small_preview_height' => null,
-                    'video_small_preview_width' => null,
-                    'video_small_preview_content_length' => null,
-                    'video_small_preview_content_type' => null,
-                ],
-                8 => [
-                    'id' => 283819010,
-                    'title' => 'Frog surfacing in pond',
-                    'creator_id' => 202786143,
-                    'creator_name' => 'bigemrg',
-                    'creation_date' => '2019-08-13 22:39:55',
-                    'country_name' => 'United Kingdom of Great Britain and Northern Ireland',
-                    'thumbnail_url' => null,
-                    'thumbnail_html_tag' => null,
-                    'thumbnail_width' => null,
-                    'thumbnail_height' => null,
-                    'thumbnail_110_url' => null,
-                    'thumbnail_110_width' => null,
-                    'thumbnail_110_height' => null,
-                    'thumbnail_160_url' => null,
-                    'thumbnail_160_width' => null,
-                    'thumbnail_160_height' => null,
-                    'thumbnail_220_url' => null,
-                    'thumbnail_220_width' => null,
-                    'thumbnail_220_height' => null,
-                    'thumbnail_240_url' => 'https://t3.ftcdn.net/jpg/02/83/81/90/240_F_283819010_OqAFaLciWzYbU2iSy2TbMIMrAvMJUKFi.jpg',
-                    'thumbnail_240_width' => null,
-                    'thumbnail_240_height' => null,
-                    'thumbnail_500_url' => 'https://as1.ftcdn.net/jpg/02/83/81/90/500_F_283819010_OqAFaLciWzYbU2iSy2TbMIMrAvMJUKFi.jpg',
-                    'thumbnail_500_width' => null,
-                    'thumbnail_500_height' => null,
-                    'thumbnail_1000_url' => null,
-                    'thumbnail_1000_width' => null,
-                    'thumbnail_1000_height' => null,
-                    'media_type_id' => 1,
-                    'width' => 2500,
-                    'height' => 2500,
-                    'is_licensed' => null,
-                    'comp_url' => 'https://stock.adobe.com/Rest/Libraries/Watermarked/Download/283819010/2',
-                    'comp_width' => null,
-                    'comp_height' => null,
-                    'nb_views' => null,
-                    'nb_downloads' => null,
-                    'category' => [
-                        'id' => 1,
-                        'name' => 'Animals',
-                        'link' => null,
-                    ],
-                    'keywords' => [
-                        0 => [
-                            'name' => 'frog',
-                        ],
-                        1 => [
-                            'name' => 'duck pond',
-                        ],
-                        2 => [
-                            'name' => 'wildlife',
-                        ],
-                        3 => [
-                            'name' => 'water',
-                        ],
-                        4 => [
-                            'name' => 'head',
-                        ],
-                        5 => [
-                            'name' => 'surface',
-                        ],
-                        6 => [
-                            'name' => 'look',
-                        ],
-                        7 => [
-                            'name' => 'eye',
-                        ],
-                        8 => [
-                            'name' => 'brown',
-                        ],
-                        9 => [
-                            'name' => 'safety',
-                        ],
-                        10 => [
-                            'name' => 'outdoors',
-                        ],
-                        11 => [
-                            'name' => 'day',
-                        ],
-                        12 => [
-                            'name' => '1',
-                        ],
-                        13 => [
-                            'name' => 'animal',
-                        ],
-                        14 => [
-                            'name' => 'no people',
-                        ],
-                        15 => [
-                            'name' => 'nobody',
-                        ],
-                    ],
-                    'has_releases' => null,
-                    'asset_type_id' => null,
-                    'vector_type' => null,
-                    'content_type' => 'image/jpeg',
-                    'framerate' => null,
-                    'duration' => null,
-                    'stock_id' => null,
-                    'comps' => null,
-                    'details_url' => 'https://stock.adobe.com/283819010?as_channel=affiliate&as_source=api&as_content=7f387d439e4f4e64ae42060fc571c456',
-                    'template_type_id' => null,
-                    'template_category_ids' => null,
-                    'marketing_text' => null,
-                    'description' => null,
-                    'size_bytes' => null,
-                    'premium_level_id' => 0,
-                    'is_premium' => null,
-                    'licenses' => null,
-                    'video_preview_url' => null,
-                    'VideoPreviewHeight' => null,
-                    'video_preview_width' => null,
-                    'video_preview_content_length' => null,
-                    'video_preview_content_type' => null,
-                    'video_small_preview_url' => null,
-                    'video_small_preview_height' => null,
-                    'video_small_preview_width' => null,
-                    'video_small_preview_content_length' => null,
-                    'video_small_preview_content_type' => null,
-                ],
-                9 => [
-                    'id' => 249315530,
-                    'title' => 'Close up of woman with rainbow patterned band on wrist',
-                    'creator_id' => 200982272,
-                    'creator_name' => 'nito',
-                    'creation_date' => '2019-02-14 14:46:29',
-                    'country_name' => 'Spain',
-                    'thumbnail_url' => null,
-                    'thumbnail_html_tag' => null,
-                    'thumbnail_width' => null,
-                    'thumbnail_height' => null,
-                    'thumbnail_110_url' => null,
-                    'thumbnail_110_width' => null,
-                    'thumbnail_110_height' => null,
-                    'thumbnail_160_url' => null,
-                    'thumbnail_160_width' => null,
-                    'thumbnail_160_height' => null,
-                    'thumbnail_220_url' => null,
-                    'thumbnail_220_width' => null,
-                    'thumbnail_220_height' => null,
-                    'thumbnail_240_url' => 'https://t4.ftcdn.net/jpg/02/49/31/55/240_F_249315530_xEA9A74hZlG5j8t1Aomhqqq82kBhjuIW.jpg',
-                    'thumbnail_240_width' => null,
-                    'thumbnail_240_height' => null,
-                    'thumbnail_500_url' => 'https://as2.ftcdn.net/jpg/02/49/31/55/500_F_249315530_xEA9A74hZlG5j8t1Aomhqqq82kBhjuIW.jpg',
-                    'thumbnail_500_width' => null,
-                    'thumbnail_500_height' => null,
-                    'thumbnail_1000_url' => null,
-                    'thumbnail_1000_width' => null,
-                    'thumbnail_1000_height' => null,
-                    'media_type_id' => 1,
-                    'width' => 3470,
-                    'height' => 5200,
-                    'is_licensed' => null,
-                    'comp_url' => 'https://stock.adobe.com/Rest/Libraries/Watermarked/Download/249315530/2',
-                    'comp_width' => null,
-                    'comp_height' => null,
-                    'nb_views' => null,
-                    'nb_downloads' => null,
-                    'category' => [
-                        'id' => 695,
-                        'name' => 'People',
-                        'link' => null,
-                    ],
-                    'keywords' => [
-                        0 => [
-                            'name' => 'woman',
-                        ],
-                        1 => [
-                            'name' => 'adult',
-                        ],
-                        2 => [
-                            'name' => '1',
-                        ],
-                        3 => [
-                            'name' => 'alone',
-                        ],
-                        4 => [
-                            'name' => 'hand',
-                        ],
-                        5 => [
-                            'name' => 'wrist',
-                        ],
-                        6 => [
-                            'name' => 'rainbow',
-                        ],
-                        7 => [
-                            'name' => 'lesbian',
-                        ],
-                        8 => [
-                            'name' => 'identity',
-                        ],
-                        9 => [
-                            'name' => 'striped',
-                        ],
-                        10 => [
-                            'name' => 'pattern',
-                        ],
-                        11 => [
-                            'name' => 'many-coloured',
-                        ],
-                        12 => [
-                            'name' => 'colourful',
-                        ],
-                        13 => [
-                            'name' => 'band',
-                        ],
-                        14 => [
-                            'name' => 'incomplete',
-                        ],
-                        15 => [
-                            'name' => 'dress',
-                        ],
-                        16 => [
-                            'name' => 'floral',
-                        ],
-                        17 => [
-                            'name' => 'midsection',
-                        ],
-                        18 => [
-                            'name' => 'wristband',
-                        ],
-                        19 => [
-                            'name' => 'closeup',
-                        ],
-                        20 => [
-                            'name' => 'studio',
-                        ],
-                        21 => [
-                            'name' => 'indoor',
-                        ],
-                        22 => [
-                            'name' => 'day',
-                        ],
-                        23 => [
-                            'name' => 'people',
-                        ],
-                    ],
-                    'has_releases' => null,
-                    'asset_type_id' => null,
-                    'vector_type' => null,
-                    'content_type' => 'image/jpeg',
-                    'framerate' => null,
-                    'duration' => null,
-                    'stock_id' => null,
-                    'comps' => null,
-                    'details_url' => 'https://stock.adobe.com/249315530?as_channel=affiliate&as_source=api&as_content=7f387d439e4f4e64ae42060fc571c456',
-                    'template_type_id' => null,
-                    'template_category_ids' => null,
-                    'marketing_text' => null,
-                    'description' => null,
-                    'size_bytes' => null,
-                    'premium_level_id' => 0,
-                    'is_premium' => null,
-                    'licenses' => null,
-                    'video_preview_url' => null,
-                    'VideoPreviewHeight' => null,
-                    'video_preview_width' => null,
-                    'video_preview_content_length' => null,
-                    'video_preview_content_type' => null,
-                    'video_small_preview_url' => null,
-                    'video_small_preview_height' => null,
-                    'video_small_preview_width' => null,
-                    'video_small_preview_content_length' => null,
-                    'video_small_preview_content_type' => null,
-                ],
-                10 => [
-                    'id' => 301540362,
-                    'title' => 'Pensive woman in shadow',
-                    'creator_id' => 209113731,
-                    'creator_name' => 'Lais',
-                    'creation_date' => '2019-11-09 10:46:26',
-                    'country_name' => 'Brazil',
-                    'thumbnail_url' => null,
-                    'thumbnail_html_tag' => null,
-                    'thumbnail_width' => null,
-                    'thumbnail_height' => null,
-                    'thumbnail_110_url' => null,
-                    'thumbnail_110_width' => null,
-                    'thumbnail_110_height' => null,
-                    'thumbnail_160_url' => null,
-                    'thumbnail_160_width' => null,
-                    'thumbnail_160_height' => null,
-                    'thumbnail_220_url' => null,
-                    'thumbnail_220_width' => null,
-                    'thumbnail_220_height' => null,
-                    'thumbnail_240_url' => 'https://t4.ftcdn.net/jpg/03/01/54/03/240_F_301540362_Tkg0aCdh5u9NE9PEXuSGKvZKYZ09U6Xn.jpg',
-                    'thumbnail_240_width' => null,
-                    'thumbnail_240_height' => null,
-                    'thumbnail_500_url' => 'https://as2.ftcdn.net/jpg/03/01/54/03/500_F_301540362_Tkg0aCdh5u9NE9PEXuSGKvZKYZ09U6Xn.jpg',
-                    'thumbnail_500_width' => null,
-                    'thumbnail_500_height' => null,
-                    'thumbnail_1000_url' => null,
-                    'thumbnail_1000_width' => null,
-                    'thumbnail_1000_height' => null,
-                    'media_type_id' => 1,
-                    'width' => 3163,
-                    'height' => 4744,
-                    'is_licensed' => null,
-                    'comp_url' => 'https://stock.adobe.com/Rest/Libraries/Watermarked/Download/301540362/2',
-                    'comp_width' => null,
-                    'comp_height' => null,
-                    'nb_views' => null,
-                    'nb_downloads' => null,
-                    'category' => [
-                        'id' => 695,
-                        'name' => 'People',
-                        'link' => null,
-                    ],
-                    'keywords' => [
-                        0 => [
-                            'name' => 'woman',
-                        ],
-                        1 => [
-                            'name' => 'adult',
-                        ],
-                        2 => [
-                            'name' => 'young',
-                        ],
-                        3 => [
-                            'name' => 'alone',
-                        ],
-                        4 => [
-                            'name' => 'shadow',
-                        ],
-                        5 => [
-                            'name' => 'sunlight',
-                        ],
-                        6 => [
-                            'name' => 'pattern',
-                        ],
-                        7 => [
-                            'name' => 'table',
-                        ],
-                        8 => [
-                            'name' => 'lean',
-                        ],
-                        9 => [
-                            'name' => 'eyes closed',
-                        ],
-                        10 => [
-                            'name' => 'pensive',
-                        ],
-                        11 => [
-                            'name' => 'dream',
-                        ],
-                        12 => [
-                            'name' => 'contemplation',
-                        ],
-                        13 => [
-                            'name' => 'sleepy',
-                        ],
-                        14 => [
-                            'name' => 'sad',
-                        ],
-                        15 => [
-                            'name' => 'indoor',
-                        ],
-                        16 => [
-                            'name' => 'day',
-                        ],
-                        17 => [
-                            'name' => '1',
-                        ],
-                        18 => [
-                            'name' => 'people',
-                        ],
-                    ],
-                    'has_releases' => null,
-                    'asset_type_id' => null,
-                    'vector_type' => null,
-                    'content_type' => 'image/jpeg',
-                    'framerate' => null,
-                    'duration' => null,
-                    'stock_id' => null,
-                    'comps' => null,
-                    'details_url' => 'https://stock.adobe.com/301540362?as_channel=affiliate&as_source=api&as_content=7f387d439e4f4e64ae42060fc571c456',
-                    'template_type_id' => null,
-                    'template_category_ids' => null,
-                    'marketing_text' => null,
-                    'description' => null,
-                    'size_bytes' => null,
-                    'premium_level_id' => 0,
-                    'is_premium' => null,
-                    'licenses' => null,
-                    'video_preview_url' => null,
-                    'VideoPreviewHeight' => null,
-                    'video_preview_width' => null,
-                    'video_preview_content_length' => null,
-                    'video_preview_content_type' => null,
-                    'video_small_preview_url' => null,
-                    'video_small_preview_height' => null,
-                    'video_small_preview_width' => null,
-                    'video_small_preview_content_length' => null,
-                    'video_small_preview_content_type' => null,
-                ],
-                11 => [
-                    'id' => 303644064,
-                    'title' => 'Young woman sunbathing with closed eyes on swim ring in swimming pool',
-                    'creator_id' => 206713618,
-                    'creator_name' => 'LIGHTFIELD STUDIOS',
-                    'creation_date' => '2019-11-18 19:29:53',
-                    'country_name' => 'United States of America',
-                    'thumbnail_url' => null,
-                    'thumbnail_html_tag' => null,
-                    'thumbnail_width' => null,
-                    'thumbnail_height' => null,
-                    'thumbnail_110_url' => null,
-                    'thumbnail_110_width' => null,
-                    'thumbnail_110_height' => null,
-                    'thumbnail_160_url' => null,
-                    'thumbnail_160_width' => null,
-                    'thumbnail_160_height' => null,
-                    'thumbnail_220_url' => null,
-                    'thumbnail_220_width' => null,
-                    'thumbnail_220_height' => null,
-                    'thumbnail_240_url' => 'https://t3.ftcdn.net/jpg/03/03/64/40/240_F_303644064_T5o26ZZGWxKq80rFTWhigotG033CHwKa.jpg',
-                    'thumbnail_240_width' => null,
-                    'thumbnail_240_height' => null,
-                    'thumbnail_500_url' => 'https://as1.ftcdn.net/jpg/03/03/64/40/500_F_303644064_T5o26ZZGWxKq80rFTWhigotG033CHwKa.jpg',
-                    'thumbnail_500_width' => null,
-                    'thumbnail_500_height' => null,
-                    'thumbnail_1000_url' => null,
-                    'thumbnail_1000_width' => null,
-                    'thumbnail_1000_height' => null,
-                    'media_type_id' => 1,
-                    'width' => 4912,
-                    'height' => 7360,
-                    'is_licensed' => null,
-                    'comp_url' => 'https://stock.adobe.com/Rest/Libraries/Watermarked/Download/303644064/2',
-                    'comp_width' => null,
-                    'comp_height' => null,
-                    'nb_views' => null,
-                    'nb_downloads' => null,
-                    'category' => [
-                        'id' => 596,
-                        'name' => 'Landscapes',
-                        'link' => null,
-                    ],
-                    'keywords' => [
-                        0 => [
-                            'name' => 'woman',
-                        ],
-                        1 => [
-                            'name' => 'adult',
-                        ],
-                        2 => [
-                            'name' => 'young',
-                        ],
-                        3 => [
-                            'name' => 'caucasian',
-                        ],
-                        4 => [
-                            'name' => 'alone',
-                        ],
-                        5 => [
-                            'name' => 'inflatable',
-                        ],
-                        6 => [
-                            'name' => 'ring',
-                        ],
-                        7 => [
-                            'name' => 'swimming pool',
-                        ],
-                        8 => [
-                            'name' => 'water',
-                        ],
-                        9 => [
-                            'name' => 'swimwear',
-                        ],
-                        10 => [
-                            'name' => 'to sit',
-                        ],
-                        11 => [
-                            'name' => 'sunbath',
-                        ],
-                        12 => [
-                            'name' => 'resort',
-                        ],
-                        13 => [
-                            'name' => 'brunette',
-                        ],
-                        14 => [
-                            'name' => 'long hair',
-                        ],
-                        15 => [
-                            'name' => 'beautiful',
-                        ],
-                        16 => [
-                            'name' => 'sunny',
-                        ],
-                        17 => [
-                            'name' => 'summer',
-                        ],
-                        18 => [
-                            'name' => 'vacation',
-                        ],
-                        19 => [
-                            'name' => 'relaxation',
-                        ],
-                        20 => [
-                            'name' => 'leisure',
-                        ],
-                        21 => [
-                            'name' => 'lifestyle',
-                        ],
-                        22 => [
-                            'name' => 'outdoors',
-                        ],
-                        23 => [
-                            'name' => 'day',
-                        ],
-                        24 => [
-                            'name' => '1',
-                        ],
-                        25 => [
-                            'name' => 'people',
-                        ],
-                    ],
-                    'has_releases' => null,
-                    'asset_type_id' => null,
-                    'vector_type' => null,
-                    'content_type' => 'image/jpeg',
-                    'framerate' => null,
-                    'duration' => null,
-                    'stock_id' => null,
-                    'comps' => null,
-                    'details_url' => 'https://stock.adobe.com/303644064?as_channel=affiliate&as_source=api&as_content=7f387d439e4f4e64ae42060fc571c456',
-                    'template_type_id' => null,
-                    'template_category_ids' => null,
-                    'marketing_text' => null,
-                    'description' => null,
-                    'size_bytes' => null,
-                    'premium_level_id' => 0,
-                    'is_premium' => null,
-                    'licenses' => null,
-                    'video_preview_url' => null,
-                    'VideoPreviewHeight' => null,
-                    'video_preview_width' => null,
-                    'video_preview_content_length' => null,
-                    'video_preview_content_type' => null,
-                    'video_small_preview_url' => null,
-                    'video_small_preview_height' => null,
-                    'video_small_preview_width' => null,
-                    'video_small_preview_content_length' => null,
-                    'video_small_preview_content_type' => null,
-                ],
-                12 => [
-                    'id' => 306816678,
-                    'title' => 'Rear view of woman in long white dress on swing in landscape',
-                    'creator_id' => 206123944,
-                    'creator_name' => 'Oleg Breslavtsev',
-                    'creation_date' => '2019-12-02 04:00:28',
-                    'country_name' => 'Russian Federation',
-                    'thumbnail_url' => null,
-                    'thumbnail_html_tag' => null,
-                    'thumbnail_width' => null,
-                    'thumbnail_height' => null,
-                    'thumbnail_110_url' => null,
-                    'thumbnail_110_width' => null,
-                    'thumbnail_110_height' => null,
-                    'thumbnail_160_url' => null,
-                    'thumbnail_160_width' => null,
-                    'thumbnail_160_height' => null,
-                    'thumbnail_220_url' => null,
-                    'thumbnail_220_width' => null,
-                    'thumbnail_220_height' => null,
-                    'thumbnail_240_url' => 'https://t3.ftcdn.net/jpg/03/06/81/66/240_F_306816678_gh0NHxqrTnsy4VksmHaIu81k9tATUqp1.jpg',
-                    'thumbnail_240_width' => null,
-                    'thumbnail_240_height' => null,
-                    'thumbnail_500_url' => 'https://as1.ftcdn.net/jpg/03/06/81/66/500_F_306816678_gh0NHxqrTnsy4VksmHaIu81k9tATUqp1.jpg',
-                    'thumbnail_500_width' => null,
-                    'thumbnail_500_height' => null,
-                    'thumbnail_1000_url' => null,
-                    'thumbnail_1000_width' => null,
-                    'thumbnail_1000_height' => null,
-                    'media_type_id' => 1,
-                    'width' => 4480,
-                    'height' => 6720,
-                    'is_licensed' => null,
-                    'comp_url' => 'https://stock.adobe.com/Rest/Libraries/Watermarked/Download/306816678/2',
-                    'comp_width' => null,
-                    'comp_height' => null,
-                    'nb_views' => null,
-                    'nb_downloads' => null,
-                    'category' => [
-                        'id' => 695,
-                        'name' => 'People',
-                        'link' => null,
-                    ],
-                    'keywords' => [
-                        0 => [
-                            'name' => 'woman',
-                        ],
-                        1 => [
-                            'name' => 'adult',
-                        ],
-                        2 => [
-                            'name' => 'young',
-                        ],
-                        3 => [
-                            'name' => '1',
-                        ],
-                        4 => [
-                            'name' => 'alone',
-                        ],
-                        5 => [
-                            'name' => 'swing',
-                        ],
-                        6 => [
-                            'name' => 'tree',
-                        ],
-                        7 => [
-                            'name' => 'dress',
-                        ],
-                        8 => [
-                            'name' => 'long',
-                        ],
-                        9 => [
-                            'name' => 'white',
-                        ],
-                        10 => [
-                            'name' => 'blond',
-                        ],
-                        11 => [
-                            'name' => 'long hair',
-                        ],
-                        12 => [
-                            'name' => 'sea',
-                        ],
-                        13 => [
-                            'name' => 'sunlight',
-                        ],
-                        14 => [
-                            'name' => 'landscape',
-                        ],
-                        15 => [
-                            'name' => 'scenic',
-                        ],
-                        16 => [
-                            'name' => 'cliff',
-                        ],
-                        17 => [
-                            'name' => 'dream',
-                        ],
-                        18 => [
-                            'name' => 'harmony',
-                        ],
-                        19 => [
-                            'name' => 'wanderlust',
-                        ],
-                        20 => [
-                            'name' => 'romantic',
-                        ],
-                        21 => [
-                            'name' => 'travel',
-                        ],
-                        22 => [
-                            'name' => 'tropical',
-                        ],
-                        23 => [
-                            'name' => 'idylli',
-                        ],
-                        24 => [
-                            'name' => 'bali',
-                        ],
-                        25 => [
-                            'name' => 'outdoors',
-                        ],
-                        26 => [
-                            'name' => 'day',
-                        ],
-                        27 => [
-                            'name' => 'lifestyle',
-                        ],
-                        28 => [
-                            'name' => 'people',
-                        ],
-                    ],
-                    'has_releases' => null,
-                    'asset_type_id' => null,
-                    'vector_type' => null,
-                    'content_type' => 'image/jpeg',
-                    'framerate' => null,
-                    'duration' => null,
-                    'stock_id' => null,
-                    'comps' => null,
-                    'details_url' => 'https://stock.adobe.com/306816678?as_channel=affiliate&as_source=api&as_content=7f387d439e4f4e64ae42060fc571c456',
-                    'template_type_id' => null,
-                    'template_category_ids' => null,
-                    'marketing_text' => null,
-                    'description' => null,
-                    'size_bytes' => null,
-                    'premium_level_id' => 0,
-                    'is_premium' => null,
-                    'licenses' => null,
-                    'video_preview_url' => null,
-                    'VideoPreviewHeight' => null,
-                    'video_preview_width' => null,
-                    'video_preview_content_length' => null,
-                    'video_preview_content_type' => null,
-                    'video_small_preview_url' => null,
-                    'video_small_preview_height' => null,
-                    'video_small_preview_width' => null,
-                    'video_small_preview_content_length' => null,
-                    'video_small_preview_content_type' => null,
-                ],
-                13 => [
-                    'id' => 275348326,
-                    'title' => 'Portrait of happy family with dog sitting on floor in bedroom',
-                    'creator_id' => 205688239,
-                    'creator_name' => 'natalialeb',
-                    'creation_date' => '2019-06-26 22:37:24',
-                    'country_name' => 'Russian Federation',
-                    'thumbnail_url' => null,
-                    'thumbnail_html_tag' => null,
-                    'thumbnail_width' => null,
-                    'thumbnail_height' => null,
-                    'thumbnail_110_url' => null,
-                    'thumbnail_110_width' => null,
-                    'thumbnail_110_height' => null,
-                    'thumbnail_160_url' => null,
-                    'thumbnail_160_width' => null,
-                    'thumbnail_160_height' => null,
-                    'thumbnail_220_url' => null,
-                    'thumbnail_220_width' => null,
-                    'thumbnail_220_height' => null,
-                    'thumbnail_240_url' => 'https://t4.ftcdn.net/jpg/02/75/34/83/240_F_275348326_I2A6YQg2NyZ0ROLBdl96N8h6Q0NltqT2.jpg',
-                    'thumbnail_240_width' => null,
-                    'thumbnail_240_height' => null,
-                    'thumbnail_500_url' => 'https://as2.ftcdn.net/jpg/02/75/34/83/500_F_275348326_I2A6YQg2NyZ0ROLBdl96N8h6Q0NltqT2.jpg',
-                    'thumbnail_500_width' => null,
-                    'thumbnail_500_height' => null,
-                    'thumbnail_1000_url' => null,
-                    'thumbnail_1000_width' => null,
-                    'thumbnail_1000_height' => null,
-                    'media_type_id' => 1,
-                    'width' => 5760,
-                    'height' => 3840,
-                    'is_licensed' => null,
-                    'comp_url' => 'https://stock.adobe.com/Rest/Libraries/Watermarked/Download/275348326/2',
-                    'comp_width' => null,
-                    'comp_height' => null,
-                    'nb_views' => null,
-                    'nb_downloads' => null,
-                    'category' => [
-                        'id' => 695,
-                        'name' => 'People',
-                        'link' => null,
-                    ],
-                    'keywords' => [
-                        0 => [
-                            'name' => 'woman',
-                        ],
-                        1 => [
-                            'name' => 'man',
-                        ],
-                        2 => [
-                            'name' => 'girl',
-                        ],
-                        3 => [
-                            'name' => 'adult',
-                        ],
-                        4 => [
-                            'name' => 'children',
-                        ],
-                        5 => [
-                            'name' => 'young',
-                        ],
-                        6 => [
-                            'name' => 'toddler',
-                        ],
-                        7 => [
-                            'name' => 'elementary',
-                        ],
-                        8 => [
-                            'name' => 'caucasian',
-                        ],
-                        9 => [
-                            'name' => 'mother',
-                        ],
-                        10 => [
-                            'name' => 'father',
-                        ],
-                        11 => [
-                            'name' => 'daughter',
-                        ],
-                        12 => [
-                            'name' => 'family',
-                        ],
-                        13 => [
-                            'name' => 'smile',
-                        ],
-                        14 => [
-                            'name' => 'bedchamber',
-                        ],
-                        15 => [
-                            'name' => 'down\'s syndrome',
-                        ],
-                        16 => [
-                            'name' => 'disability',
-                        ],
-                        17 => [
-                            'name' => 'to sit',
-                        ],
-                        18 => [
-                            'name' => 'hugs',
-                        ],
-                        19 => [
-                            'name' => 'blond',
-                        ],
-                        20 => [
-                            'name' => 'casual attire',
-                        ],
-                        21 => [
-                            'name' => 'dog',
-                        ],
-                        22 => [
-                            'name' => 'wave',
-                        ],
-                        23 => [
-                            'name' => 'cheerful',
-                        ],
-                        24 => [
-                            'name' => 'happiness',
-                        ],
-                        25 => [
-                            'name' => 'togetherness',
-                        ],
-                        26 => [
-                            'name' => 'domestic',
-                        ],
-                        27 => [
-                            'name' => 'lifestyle',
-                        ],
-                        28 => [
-                            'name' => 'pet',
-                        ],
-                        29 => [
-                            'name' => 'portrait',
-                        ],
-                        30 => [
-                            'name' => 'indoor',
-                        ],
-                        31 => [
-                            'name' => 'day',
-                        ],
-                        32 => [
-                            'name' => 'group',
-                        ],
-                        33 => [
-                            'name' => '4',
-                        ],
-                        34 => [
-                            'name' => 'people',
-                        ],
-                        35 => [
-                            'name' => '1',
-                        ],
-                        36 => [
-                            'name' => 'animal',
-                        ],
-                    ],
-                    'has_releases' => null,
-                    'asset_type_id' => null,
-                    'vector_type' => null,
-                    'content_type' => 'image/jpeg',
-                    'framerate' => null,
-                    'duration' => null,
-                    'stock_id' => null,
-                    'comps' => null,
-                    'details_url' => 'https://stock.adobe.com/275348326?as_channel=affiliate&as_source=api&as_content=7f387d439e4f4e64ae42060fc571c456',
-                    'template_type_id' => null,
-                    'template_category_ids' => null,
-                    'marketing_text' => null,
-                    'description' => null,
-                    'size_bytes' => null,
-                    'premium_level_id' => 0,
-                    'is_premium' => null,
-                    'licenses' => null,
-                    'video_preview_url' => null,
-                    'VideoPreviewHeight' => null,
-                    'video_preview_width' => null,
-                    'video_preview_content_length' => null,
-                    'video_preview_content_type' => null,
-                    'video_small_preview_url' => null,
-                    'video_small_preview_height' => null,
-                    'video_small_preview_width' => null,
-                    'video_small_preview_content_length' => null,
-                    'video_small_preview_content_type' => null,
-                ],
-                14 => [
-                    'id' => 274053113,
-                    'title' => 'Overhead view of father and son playing',
-                    'creator_id' => 205198462,
-                    'creator_name' => 'StockRocket',
-                    'creation_date' => '2019-06-19 14:44:39',
-                    'country_name' => 'Serbia',
-                    'thumbnail_url' => null,
-                    'thumbnail_html_tag' => null,
-                    'thumbnail_width' => null,
-                    'thumbnail_height' => null,
-                    'thumbnail_110_url' => null,
-                    'thumbnail_110_width' => null,
-                    'thumbnail_110_height' => null,
-                    'thumbnail_160_url' => null,
-                    'thumbnail_160_width' => null,
-                    'thumbnail_160_height' => null,
-                    'thumbnail_220_url' => null,
-                    'thumbnail_220_width' => null,
-                    'thumbnail_220_height' => null,
-                    'thumbnail_240_url' => 'https://t4.ftcdn.net/jpg/02/74/05/31/240_F_274053113_jUjXnHfayfGk4ybdxKYlMeq65tTebb2B.jpg',
-                    'thumbnail_240_width' => null,
-                    'thumbnail_240_height' => null,
-                    'thumbnail_500_url' => 'https://as2.ftcdn.net/jpg/02/74/05/31/500_F_274053113_jUjXnHfayfGk4ybdxKYlMeq65tTebb2B.jpg',
-                    'thumbnail_500_width' => null,
-                    'thumbnail_500_height' => null,
-                    'thumbnail_1000_url' => null,
-                    'thumbnail_1000_width' => null,
-                    'thumbnail_1000_height' => null,
-                    'media_type_id' => 1,
-                    'width' => 4242,
-                    'height' => 4341,
-                    'is_licensed' => null,
-                    'comp_url' => 'https://stock.adobe.com/Rest/Libraries/Watermarked/Download/274053113/2',
-                    'comp_width' => null,
-                    'comp_height' => null,
-                    'nb_views' => null,
-                    'nb_downloads' => null,
-                    'category' => [
-                        'id' => 695,
-                        'name' => 'People',
-                        'link' => null,
-                    ],
-                    'keywords' => [
-                        0 => [
-                            'name' => 'man',
-                        ],
-                        1 => [
-                            'name' => 'boy',
-                        ],
-                        2 => [
-                            'name' => 'adult',
-                        ],
-                        3 => [
-                            'name' => 'children',
-                        ],
-                        4 => [
-                            'name' => 'mid',
-                        ],
-                        5 => [
-                            'name' => 'elementary',
-                        ],
-                        6 => [
-                            'name' => 'caucasian',
-                        ],
-                        7 => [
-                            'name' => 'pair',
-                        ],
-                        8 => [
-                            'name' => 'father',
-                        ],
-                        9 => [
-                            'name' => 'son',
-                        ],
-                        10 => [
-                            'name' => 'lay',
-                        ],
-                        11 => [
-                            'name' => 'smile',
-                        ],
-                        12 => [
-                            'name' => 'play',
-                        ],
-                        13 => [
-                            'name' => 'casual attire',
-                        ],
-                        14 => [
-                            'name' => 't-shirt',
-                        ],
-                        15 => [
-                            'name' => 'rug',
-                        ],
-                        16 => [
-                            'name' => 'pattern',
-                        ],
-                        17 => [
-                            'name' => 'hugs',
-                        ],
-                        18 => [
-                            'name' => 'togetherness',
-                        ],
-                        19 => [
-                            'name' => 'bond',
-                        ],
-                        20 => [
-                            'name' => 'enjoyment',
-                        ],
-                        21 => [
-                            'name' => 'love',
-                        ],
-                        22 => [
-                            'name' => 'indoor',
-                        ],
-                        23 => [
-                            'name' => 'day',
-                        ],
-                        24 => [
-                            'name' => '2',
-                        ],
-                        25 => [
-                            'name' => 'people',
-                        ],
-                    ],
-                    'has_releases' => null,
-                    'asset_type_id' => null,
-                    'vector_type' => null,
-                    'content_type' => 'image/jpeg',
-                    'framerate' => null,
-                    'duration' => null,
-                    'stock_id' => null,
-                    'comps' => null,
-                    'details_url' => 'https://stock.adobe.com/274053113?as_channel=affiliate&as_source=api&as_content=7f387d439e4f4e64ae42060fc571c456',
-                    'template_type_id' => null,
-                    'template_category_ids' => null,
-                    'marketing_text' => null,
-                    'description' => null,
-                    'size_bytes' => null,
-                    'premium_level_id' => 0,
-                    'is_premium' => null,
-                    'licenses' => null,
-                    'video_preview_url' => null,
-                    'VideoPreviewHeight' => null,
-                    'video_preview_width' => null,
-                    'video_preview_content_length' => null,
-                    'video_preview_content_type' => null,
-                    'video_small_preview_url' => null,
-                    'video_small_preview_height' => null,
-                    'video_small_preview_width' => null,
-                    'video_small_preview_content_length' => null,
-                    'video_small_preview_content_type' => null,
-                ],
-                15 => [
-                    'id' => 250445367,
-                    'title' => 'Two girls on bicycles on street',
-                    'creator_id' => 224608,
-                    'creator_name' => 'Jacob Lund ',
-                    'creation_date' => '2019-02-20 15:33:05',
-                    'country_name' => 'Denmark',
-                    'thumbnail_url' => null,
-                    'thumbnail_html_tag' => null,
-                    'thumbnail_width' => null,
-                    'thumbnail_height' => null,
-                    'thumbnail_110_url' => null,
-                    'thumbnail_110_width' => null,
-                    'thumbnail_110_height' => null,
-                    'thumbnail_160_url' => null,
-                    'thumbnail_160_width' => null,
-                    'thumbnail_160_height' => null,
-                    'thumbnail_220_url' => null,
-                    'thumbnail_220_width' => null,
-                    'thumbnail_220_height' => null,
-                    'thumbnail_240_url' => 'https://t4.ftcdn.net/jpg/02/50/44/53/240_F_250445367_vLqBRyGxAWVvajz1HjLv4LHS9rM0NxA9.jpg',
-                    'thumbnail_240_width' => null,
-                    'thumbnail_240_height' => null,
-                    'thumbnail_500_url' => 'https://as2.ftcdn.net/jpg/02/50/44/53/500_F_250445367_vLqBRyGxAWVvajz1HjLv4LHS9rM0NxA9.jpg',
-                    'thumbnail_500_width' => null,
-                    'thumbnail_500_height' => null,
-                    'thumbnail_1000_url' => null,
-                    'thumbnail_1000_width' => null,
-                    'thumbnail_1000_height' => null,
-                    'media_type_id' => 1,
-                    'width' => 7590,
-                    'height' => 5060,
-                    'is_licensed' => null,
-                    'comp_url' => 'https://stock.adobe.com/Rest/Libraries/Watermarked/Download/250445367/2',
-                    'comp_width' => null,
-                    'comp_height' => null,
-                    'nb_views' => null,
-                    'nb_downloads' => null,
-                    'category' => [
-                        'id' => 695,
-                        'name' => 'People',
-                        'link' => null,
-                    ],
-                    'keywords' => [
-                        0 => [
-                            'name' => 'girl',
-                        ],
-                        1 => [
-                            'name' => 'children',
-                        ],
-                        2 => [
-                            'name' => 'teenage',
-                        ],
-                        3 => [
-                            'name' => 'pair',
-                        ],
-                        4 => [
-                            'name' => 'bicycle',
-                        ],
-                        5 => [
-                            'name' => 'street',
-                        ],
-                        6 => [
-                            'name' => 'smile',
-                        ],
-                        7 => [
-                            'name' => 'blond',
-                        ],
-                        8 => [
-                            'name' => 'long hair',
-                        ],
-                        9 => [
-                            'name' => 'casual attire',
-                        ],
-                        10 => [
-                            'name' => 't-shirt',
-                        ],
-                        11 => [
-                            'name' => 'short',
-                        ],
-                        12 => [
-                            'name' => 'recreation',
-                        ],
-                        13 => [
-                            'name' => 'friendship',
-                        ],
-                        14 => [
-                            'name' => 'leisure',
-                        ],
-                        15 => [
-                            'name' => 'lifestyle',
-                        ],
-                        16 => [
-                            'name' => 'outdoors',
-                        ],
-                        17 => [
-                            'name' => 'day',
-                        ],
-                        18 => [
-                            'name' => '2',
-                        ],
-                        19 => [
-                            'name' => 'people',
-                        ],
-                    ],
-                    'has_releases' => null,
-                    'asset_type_id' => null,
-                    'vector_type' => null,
-                    'content_type' => 'image/jpeg',
-                    'framerate' => null,
-                    'duration' => null,
-                    'stock_id' => null,
-                    'comps' => null,
-                    'details_url' => 'https://stock.adobe.com/250445367?as_channel=affiliate&as_source=api&as_content=7f387d439e4f4e64ae42060fc571c456',
-                    'template_type_id' => null,
-                    'template_category_ids' => null,
-                    'marketing_text' => null,
-                    'description' => null,
-                    'size_bytes' => null,
-                    'premium_level_id' => 0,
-                    'is_premium' => null,
-                    'licenses' => null,
-                    'video_preview_url' => null,
-                    'VideoPreviewHeight' => null,
-                    'video_preview_width' => null,
-                    'video_preview_content_length' => null,
-                    'video_preview_content_type' => null,
-                    'video_small_preview_url' => null,
-                    'video_small_preview_height' => null,
-                    'video_small_preview_width' => null,
-                    'video_small_preview_content_length' => null,
-                    'video_small_preview_content_type' => null,
-                ],
-                16 => [
-                    'id' => 281266321,
-                    'title' => 'Smiling man on ladder in greenhouse with tomatoes',
-                    'creator_id' => 204337425,
-                    'creator_name' => 'rh2010',
-                    'creation_date' => '2019-07-30 16:45:22',
-                    'country_name' => 'Ukraine',
-                    'thumbnail_url' => null,
-                    'thumbnail_html_tag' => null,
-                    'thumbnail_width' => null,
-                    'thumbnail_height' => null,
-                    'thumbnail_110_url' => null,
-                    'thumbnail_110_width' => null,
-                    'thumbnail_110_height' => null,
-                    'thumbnail_160_url' => null,
-                    'thumbnail_160_width' => null,
-                    'thumbnail_160_height' => null,
-                    'thumbnail_220_url' => null,
-                    'thumbnail_220_width' => null,
-                    'thumbnail_220_height' => null,
-                    'thumbnail_240_url' => 'https://t4.ftcdn.net/jpg/02/81/26/63/240_F_281266321_Y20WMaFGtspOjoqsE3gUi11p5I4N3CcS.jpg',
-                    'thumbnail_240_width' => null,
-                    'thumbnail_240_height' => null,
-                    'thumbnail_500_url' => 'https://as2.ftcdn.net/jpg/02/81/26/63/500_F_281266321_Y20WMaFGtspOjoqsE3gUi11p5I4N3CcS.jpg',
-                    'thumbnail_500_width' => null,
-                    'thumbnail_500_height' => null,
-                    'thumbnail_1000_url' => null,
-                    'thumbnail_1000_width' => null,
-                    'thumbnail_1000_height' => null,
-                    'media_type_id' => 1,
-                    'width' => 7360,
-                    'height' => 4912,
-                    'is_licensed' => null,
-                    'comp_url' => 'https://stock.adobe.com/Rest/Libraries/Watermarked/Download/281266321/2',
-                    'comp_width' => null,
-                    'comp_height' => null,
-                    'nb_views' => null,
-                    'nb_downloads' => null,
-                    'category' => [
-                        'id' => 782,
-                        'name' => 'Plants and Flowers',
-                        'link' => null,
-                    ],
-                    'keywords' => [
-                        0 => [
-                            'name' => 'man',
-                        ],
-                        1 => [
-                            'name' => 'adult',
-                        ],
-                        2 => [
-                            'name' => 'senior',
-                        ],
-                        3 => [
-                            'name' => 'caucasian',
-                        ],
-                        4 => [
-                            'name' => 'alone',
-                        ],
-                        5 => [
-                            'name' => 'greenhouse',
-                        ],
-                        6 => [
-                            'name' => 'tomatoes',
-                        ],
-                        7 => [
-                            'name' => 'farmer',
-                        ],
-                        8 => [
-                            'name' => 'ladder',
-                        ],
-                        9 => [
-                            'name' => 'leghorn',
-                        ],
-                        10 => [
-                            'name' => 'hat',
-                        ],
-                        11 => [
-                            'name' => 'white hair',
-                        ],
-                        12 => [
-                            'name' => 'beard',
-                        ],
-                        13 => [
-                            'name' => 'smile',
-                        ],
-                        14 => [
-                            'name' => 'apron',
-                        ],
-                        15 => [
-                            'name' => 'glove',
-                        ],
-                        16 => [
-                            'name' => 'tool',
-                        ],
-                        17 => [
-                            'name' => 'agriculture',
-                        ],
-                        18 => [
-                            'name' => 'growth',
-                        ],
-                        19 => [
-                            'name' => 'leaf',
-                        ],
-                        20 => [
-                            'name' => 'sunlight',
-                        ],
-                        21 => [
-                            'name' => 'organic',
-                        ],
-                        22 => [
-                            'name' => 'vegetable',
-                        ],
-                        23 => [
-                            'name' => 'harvest',
-                        ],
-                        24 => [
-                            'name' => 'indoor',
-                        ],
-                        25 => [
-                            'name' => 'day',
-                        ],
-                        26 => [
-                            'name' => '1',
-                        ],
-                        27 => [
-                            'name' => 'people',
-                        ],
-                    ],
-                    'has_releases' => null,
-                    'asset_type_id' => null,
-                    'vector_type' => null,
-                    'content_type' => 'image/jpeg',
-                    'framerate' => null,
-                    'duration' => null,
-                    'stock_id' => null,
-                    'comps' => null,
-                    'details_url' => 'https://stock.adobe.com/281266321?as_channel=affiliate&as_source=api&as_content=7f387d439e4f4e64ae42060fc571c456',
-                    'template_type_id' => null,
-                    'template_category_ids' => null,
-                    'marketing_text' => null,
-                    'description' => null,
-                    'size_bytes' => null,
-                    'premium_level_id' => 0,
-                    'is_premium' => null,
-                    'licenses' => null,
-                    'video_preview_url' => null,
-                    'VideoPreviewHeight' => null,
-                    'video_preview_width' => null,
-                    'video_preview_content_length' => null,
-                    'video_preview_content_type' => null,
-                    'video_small_preview_url' => null,
-                    'video_small_preview_height' => null,
-                    'video_small_preview_width' => null,
-                    'video_small_preview_content_length' => null,
-                    'video_small_preview_content_type' => null,
-                ],
-                17 => [
-                    'id' => 288163844,
-                    'title' => 'Three bottles of water with reflecting shadows on blue background.  Healthy food concept. Top view. Copy space',
-                    'creator_id' => 205065318,
-                    'creator_name' => 'thayra83',
-                    'creation_date' => '2019-09-06 15:37:44',
-                    'country_name' => 'Ukraine',
-                    'thumbnail_url' => null,
-                    'thumbnail_html_tag' => null,
-                    'thumbnail_width' => null,
-                    'thumbnail_height' => null,
-                    'thumbnail_110_url' => null,
-                    'thumbnail_110_width' => null,
-                    'thumbnail_110_height' => null,
-                    'thumbnail_160_url' => null,
-                    'thumbnail_160_width' => null,
-                    'thumbnail_160_height' => null,
-                    'thumbnail_220_url' => null,
-                    'thumbnail_220_width' => null,
-                    'thumbnail_220_height' => null,
-                    'thumbnail_240_url' => 'https://t3.ftcdn.net/jpg/02/88/16/38/240_F_288163844_rzUw0t3ZwGjL2AhpynmWkPdTscTdK512.jpg',
-                    'thumbnail_240_width' => null,
-                    'thumbnail_240_height' => null,
-                    'thumbnail_500_url' => 'https://as1.ftcdn.net/jpg/02/88/16/38/500_F_288163844_rzUw0t3ZwGjL2AhpynmWkPdTscTdK512.jpg',
-                    'thumbnail_500_width' => null,
-                    'thumbnail_500_height' => null,
-                    'thumbnail_1000_url' => null,
-                    'thumbnail_1000_width' => null,
-                    'thumbnail_1000_height' => null,
-                    'media_type_id' => 1,
-                    'width' => 6016,
-                    'height' => 4016,
-                    'is_licensed' => null,
-                    'comp_url' => 'https://stock.adobe.com/Rest/Libraries/Watermarked/Download/288163844/2',
-                    'comp_width' => null,
-                    'comp_height' => null,
-                    'nb_views' => null,
-                    'nb_downloads' => null,
-                    'category' => [
-                        'id' => 214,
-                        'name' => 'Drinks',
-                        'link' => null,
-                    ],
-                    'keywords' => [
-                        0 => [
-                            'name' => 'water',
-                        ],
-                        1 => [
-                            'name' => 'bottle',
-                        ],
-                        2 => [
-                            'name' => 'sun',
-                        ],
-                        3 => [
-                            'name' => 'glare',
-                        ],
-                        4 => [
-                            'name' => 'texture',
-                        ],
-                        5 => [
-                            'name' => 'liquid',
-                        ],
-                        6 => [
-                            'name' => 'background',
-                        ],
-                        7 => [
-                            'name' => 'blue',
-                        ],
-                        8 => [
-                            'name' => 'light',
-                        ],
-                        9 => [
-                            'name' => 'refreshing',
-                        ],
-                        10 => [
-                            'name' => 'lifestyle',
-                        ],
-                        11 => [
-                            'name' => 'thirst',
-                        ],
-                        12 => [
-                            'name' => 'product',
-                        ],
-                        13 => [
-                            'name' => 'transparent',
-                        ],
-                        14 => [
-                            'name' => 'bright',
-                        ],
-                        15 => [
-                            'name' => 'shadow',
-                        ],
-                        16 => [
-                            'name' => 'fragrance',
-                        ],
-                        17 => [
-                            'name' => 'reflex',
-                        ],
-                        18 => [
-                            'name' => 'gradient',
-                        ],
-                        19 => [
-                            'name' => 'natural',
-                        ],
-                        20 => [
-                            'name' => 'shape',
-                        ],
-                        21 => [
-                            'name' => 'drink',
-                        ],
-                        22 => [
-                            'name' => 'plastic',
-                        ],
-                        23 => [
-                            'name' => 'thirsty',
-                        ],
-                        24 => [
-                            'name' => 'hydratation',
-                        ],
-                        25 => [
-                            'name' => 'full',
-                        ],
-                        26 => [
-                            'name' => 'cold',
-                        ],
-                        27 => [
-                            'name' => 'beverage',
-                        ],
-                        28 => [
-                            'name' => 'diet',
-                        ],
-                        29 => [
-                            'name' => 'refreshment',
-                        ],
-                        30 => [
-                            'name' => 'clear',
-                        ],
-                        31 => [
-                            'name' => 'soda',
-                        ],
-                        32 => [
-                            'name' => 'diagonal',
-                        ],
-                        33 => [
-                            'name' => 'summer',
-                        ],
-                        34 => [
-                            'name' => 'abstract',
-                        ],
-                        35 => [
-                            'name' => 'sunlight',
-                        ],
-                        36 => [
-                            'name' => 'object',
-                        ],
-                        37 => [
-                            'name' => 'topview',
-                        ],
-                        38 => [
-                            'name' => 'copy space',
-                        ],
-                    ],
-                    'has_releases' => null,
-                    'asset_type_id' => null,
-                    'vector_type' => null,
-                    'content_type' => 'image/jpeg',
-                    'framerate' => null,
-                    'duration' => null,
-                    'stock_id' => null,
-                    'comps' => null,
-                    'details_url' => 'https://stock.adobe.com/288163844?as_channel=affiliate&as_source=api&as_content=7f387d439e4f4e64ae42060fc571c456',
-                    'template_type_id' => null,
-                    'template_category_ids' => null,
-                    'marketing_text' => null,
-                    'description' => null,
-                    'size_bytes' => null,
-                    'premium_level_id' => 0,
-                    'is_premium' => null,
-                    'licenses' => null,
-                    'video_preview_url' => null,
-                    'VideoPreviewHeight' => null,
-                    'video_preview_width' => null,
-                    'video_preview_content_length' => null,
-                    'video_preview_content_type' => null,
-                    'video_small_preview_url' => null,
-                    'video_small_preview_height' => null,
-                    'video_small_preview_width' => null,
-                    'video_small_preview_content_length' => null,
-                    'video_small_preview_content_type' => null,
-                ],
-                18 => [
-                    'id' => 293086210,
-                    'title' => 'Illustration of banana with flowers',
-                    'creator_id' => 205079253,
-                    'creator_name' => 'master1305',
-                    'creation_date' => '2019-10-01 10:04:24',
-                    'country_name' => 'Ukraine',
-                    'thumbnail_url' => null,
-                    'thumbnail_html_tag' => null,
-                    'thumbnail_width' => null,
-                    'thumbnail_height' => null,
-                    'thumbnail_110_url' => null,
-                    'thumbnail_110_width' => null,
-                    'thumbnail_110_height' => null,
-                    'thumbnail_160_url' => null,
-                    'thumbnail_160_width' => null,
-                    'thumbnail_160_height' => null,
-                    'thumbnail_220_url' => null,
-                    'thumbnail_220_width' => null,
-                    'thumbnail_220_height' => null,
-                    'thumbnail_240_url' => 'https://t3.ftcdn.net/jpg/02/93/08/62/240_F_293086210_EG9P4ajasXRSLGMRSHG7KVcqYkUP634B.jpg',
-                    'thumbnail_240_width' => null,
-                    'thumbnail_240_height' => null,
-                    'thumbnail_500_url' => 'https://as1.ftcdn.net/jpg/02/93/08/62/500_F_293086210_EG9P4ajasXRSLGMRSHG7KVcqYkUP634B.jpg',
-                    'thumbnail_500_width' => null,
-                    'thumbnail_500_height' => null,
-                    'thumbnail_1000_url' => null,
-                    'thumbnail_1000_width' => null,
-                    'thumbnail_1000_height' => null,
-                    'media_type_id' => 1,
-                    'width' => 3000,
-                    'height' => 3000,
-                    'is_licensed' => null,
-                    'comp_url' => 'https://stock.adobe.com/Rest/Libraries/Watermarked/Download/293086210/2',
-                    'comp_width' => null,
-                    'comp_height' => null,
-                    'nb_views' => null,
-                    'nb_downloads' => null,
-                    'category' => [
-                        'id' => 498,
-                        'name' => 'Hobbies and Leisure',
-                        'link' => null,
-                    ],
-                    'keywords' => [
-                        0 => [
-                            'name' => 'banana',
-                        ],
-                        1 => [
-                            'name' => 'flower',
-                        ],
-                        2 => [
-                            'name' => 'bunch',
-                        ],
-                        3 => [
-                            'name' => 'bouquet',
-                        ],
-                        4 => [
-                            'name' => 'fruit',
-                        ],
-                        5 => [
-                            'name' => 'abundance',
-                        ],
-                        6 => [
-                            'name' => 'concept',
-                        ],
-                        7 => [
-                            'name' => 'bizarre',
-                        ],
-                        8 => [
-                            'name' => 'creativeness',
-                        ],
-                        9 => [
-                            'name' => 'surreal',
-                        ],
-                        10 => [
-                            'name' => 'growth',
-                        ],
-                        11 => [
-                            'name' => 'food',
-                        ],
-                        12 => [
-                            'name' => 'illustration',
-                        ],
-                        13 => [
-                            'name' => 'art',
-                        ],
-                        14 => [
-                            'name' => 'graphic',
-                        ],
-                        15 => [
-                            'name' => 'no people',
-                        ],
-                        16 => [
-                            'name' => 'nobody',
-                        ],
-                    ],
-                    'has_releases' => null,
-                    'asset_type_id' => null,
-                    'vector_type' => null,
-                    'content_type' => 'image/jpeg',
-                    'framerate' => null,
-                    'duration' => null,
-                    'stock_id' => null,
-                    'comps' => null,
-                    'details_url' => 'https://stock.adobe.com/293086210?as_channel=affiliate&as_source=api&as_content=7f387d439e4f4e64ae42060fc571c456',
-                    'template_type_id' => null,
-                    'template_category_ids' => null,
-                    'marketing_text' => null,
-                    'description' => null,
-                    'size_bytes' => null,
-                    'premium_level_id' => 0,
-                    'is_premium' => null,
-                    'licenses' => null,
-                    'video_preview_url' => null,
-                    'VideoPreviewHeight' => null,
-                    'video_preview_width' => null,
-                    'video_preview_content_length' => null,
-                    'video_preview_content_type' => null,
-                    'video_small_preview_url' => null,
-                    'video_small_preview_height' => null,
-                    'video_small_preview_width' => null,
-                    'video_small_preview_content_length' => null,
-                    'video_small_preview_content_type' => null,
-                ],
-                19 => [
-                    'id' => 255967700,
-                    'title' => 'Ballerina in pink dress dancing in sunlight',
-                    'creator_id' => 206713618,
-                    'creator_name' => 'LIGHTFIELD STUDIOS',
-                    'creation_date' => '2019-03-18 13:57:32',
-                    'country_name' => 'United States of America',
-                    'thumbnail_url' => null,
-                    'thumbnail_html_tag' => null,
-                    'thumbnail_width' => null,
-                    'thumbnail_height' => null,
-                    'thumbnail_110_url' => null,
-                    'thumbnail_110_width' => null,
-                    'thumbnail_110_height' => null,
-                    'thumbnail_160_url' => null,
-                    'thumbnail_160_width' => null,
-                    'thumbnail_160_height' => null,
-                    'thumbnail_220_url' => null,
-                    'thumbnail_220_width' => null,
-                    'thumbnail_220_height' => null,
-                    'thumbnail_240_url' => 'https://t4.ftcdn.net/jpg/02/55/96/77/240_F_255967700_tkKrLKGPy0gqsInELngtp9XsZS9hSYsB.jpg',
-                    'thumbnail_240_width' => null,
-                    'thumbnail_240_height' => null,
-                    'thumbnail_500_url' => 'https://as2.ftcdn.net/jpg/02/55/96/77/500_F_255967700_tkKrLKGPy0gqsInELngtp9XsZS9hSYsB.jpg',
-                    'thumbnail_500_width' => null,
-                    'thumbnail_500_height' => null,
-                    'thumbnail_1000_url' => null,
-                    'thumbnail_1000_width' => null,
-                    'thumbnail_1000_height' => null,
-                    'media_type_id' => 1,
-                    'width' => 7360,
-                    'height' => 4912,
-                    'is_licensed' => null,
-                    'comp_url' => 'https://stock.adobe.com/Rest/Libraries/Watermarked/Download/255967700/2',
-                    'comp_width' => null,
-                    'comp_height' => null,
-                    'nb_views' => null,
-                    'nb_downloads' => null,
-                    'category' => [
-                        'id' => 695,
-                        'name' => 'People',
-                        'link' => null,
-                    ],
-                    'keywords' => [
-                        0 => [
-                            'name' => 'woman',
-                        ],
-                        1 => [
-                            'name' => 'adult',
-                        ],
-                        2 => [
-                            'name' => 'young',
-                        ],
-                        3 => [
-                            'name' => 'caucasian',
-                        ],
-                        4 => [
-                            'name' => 'alone',
-                        ],
-                        5 => [
-                            'name' => 'ballerina',
-                        ],
-                        6 => [
-                            'name' => 'ballet',
-                        ],
-                        7 => [
-                            'name' => 'dancer',
-                        ],
-                        8 => [
-                            'name' => 'dance',
-                        ],
-                        9 => [
-                            'name' => 'pink',
-                        ],
-                        10 => [
-                            'name' => 'dress',
-                        ],
-                        11 => [
-                            'name' => 'sunlight',
-                        ],
-                        12 => [
-                            'name' => 'flexibility',
-                        ],
-                        13 => [
-                            'name' => 'beautiful',
-                        ],
-                        14 => [
-                            'name' => 'grace',
-                        ],
-                        15 => [
-                            'name' => 'skill',
-                        ],
-                        16 => [
-                            'name' => 'practice',
-                        ],
-                        17 => [
-                            'name' => 'performance',
-                        ],
-                        18 => [
-                            'name' => 'indoor',
-                        ],
-                        19 => [
-                            'name' => 'day',
-                        ],
-                        20 => [
-                            'name' => '1',
-                        ],
-                        21 => [
-                            'name' => 'people',
-                        ],
-                    ],
-                    'has_releases' => null,
-                    'asset_type_id' => null,
-                    'vector_type' => null,
-                    'content_type' => 'image/jpeg',
-                    'framerate' => null,
-                    'duration' => null,
-                    'stock_id' => null,
-                    'comps' => null,
-                    'details_url' => 'https://stock.adobe.com/255967700?as_channel=affiliate&as_source=api&as_content=7f387d439e4f4e64ae42060fc571c456',
-                    'template_type_id' => null,
-                    'template_category_ids' => null,
-                    'marketing_text' => null,
-                    'description' => null,
-                    'size_bytes' => null,
-                    'premium_level_id' => 0,
-                    'is_premium' => null,
-                    'licenses' => null,
-                    'video_preview_url' => null,
-                    'VideoPreviewHeight' => null,
-                    'video_preview_width' => null,
-                    'video_preview_content_length' => null,
-                    'video_preview_content_type' => null,
-                    'video_small_preview_url' => null,
-                    'video_small_preview_height' => null,
-                    'video_small_preview_width' => null,
-                    'video_small_preview_content_length' => null,
-                    'video_small_preview_content_type' => null,
-                ],
-                20 => [
-                    'id' => 280369835,
-                    'title' => 'Happy couple playing with shopping cart in supermarket',
-                    'creator_id' => 206713618,
-                    'creator_name' => 'LIGHTFIELD STUDIOS',
-                    'creation_date' => '2019-07-25 12:48:29',
-                    'country_name' => 'United States of America',
-                    'thumbnail_url' => null,
-                    'thumbnail_html_tag' => null,
-                    'thumbnail_width' => null,
-                    'thumbnail_height' => null,
-                    'thumbnail_110_url' => null,
-                    'thumbnail_110_width' => null,
-                    'thumbnail_110_height' => null,
-                    'thumbnail_160_url' => null,
-                    'thumbnail_160_width' => null,
-                    'thumbnail_160_height' => null,
-                    'thumbnail_220_url' => null,
-                    'thumbnail_220_width' => null,
-                    'thumbnail_220_height' => null,
-                    'thumbnail_240_url' => 'https://t3.ftcdn.net/jpg/02/80/36/98/240_F_280369835_dYBmXDWoLHQGj6Vl7kcN8DaM6AgoB1Sf.jpg',
-                    'thumbnail_240_width' => null,
-                    'thumbnail_240_height' => null,
-                    'thumbnail_500_url' => 'https://as1.ftcdn.net/jpg/02/80/36/98/500_F_280369835_dYBmXDWoLHQGj6Vl7kcN8DaM6AgoB1Sf.jpg',
-                    'thumbnail_500_width' => null,
-                    'thumbnail_500_height' => null,
-                    'thumbnail_1000_url' => null,
-                    'thumbnail_1000_width' => null,
-                    'thumbnail_1000_height' => null,
-                    'media_type_id' => 1,
-                    'width' => 7360,
-                    'height' => 4912,
-                    'is_licensed' => null,
-                    'comp_url' => 'https://stock.adobe.com/Rest/Libraries/Watermarked/Download/280369835/2',
-                    'comp_width' => null,
-                    'comp_height' => null,
-                    'nb_views' => null,
-                    'nb_downloads' => null,
-                    'category' => [
-                        'id' => 695,
-                        'name' => 'People',
-                        'link' => null,
-                    ],
-                    'keywords' => [
-                        0 => [
-                            'name' => 'woman',
-                        ],
-                        1 => [
-                            'name' => 'man',
-                        ],
-                        2 => [
-                            'name' => 'adult',
-                        ],
-                        3 => [
-                            'name' => 'young',
-                        ],
-                        4 => [
-                            'name' => 'couple',
-                        ],
-                        5 => [
-                            'name' => 'shopping cart',
-                        ],
-                        6 => [
-                            'name' => 'supermarket',
-                        ],
-                        7 => [
-                            'name' => 'playful',
-                        ],
-                        8 => [
-                            'name' => 'african american',
-                        ],
-                        9 => [
-                            'name' => 'asian',
-                        ],
-                        10 => [
-                            'name' => 'casual attire',
-                        ],
-                        11 => [
-                            'name' => 'smile',
-                        ],
-                        12 => [
-                            'name' => 'laugh',
-                        ],
-                        13 => [
-                            'name' => 'jean',
-                        ],
-                        14 => [
-                            'name' => 'sunglasses',
-                        ],
-                        15 => [
-                            'name' => 'to sit',
-                        ],
-                        16 => [
-                            'name' => 'inside',
-                        ],
-                        17 => [
-                            'name' => 'push',
-                        ],
-                        18 => [
-                            'name' => 'aisle',
-                        ],
-                        19 => [
-                            'name' => 'shopping',
-                        ],
-                        20 => [
-                            'name' => 'enjoyment',
-                        ],
-                        21 => [
-                            'name' => 'lifestyle',
-                        ],
-                        22 => [
-                            'name' => 'carefree',
-                        ],
-                        23 => [
-                            'name' => 'indoor',
-                        ],
-                        24 => [
-                            'name' => 'day',
-                        ],
-                        25 => [
-                            'name' => '2',
-                        ],
-                        26 => [
-                            'name' => 'people',
-                        ],
-                    ],
-                    'has_releases' => null,
-                    'asset_type_id' => null,
-                    'vector_type' => null,
-                    'content_type' => 'image/jpeg',
-                    'framerate' => null,
-                    'duration' => null,
-                    'stock_id' => null,
-                    'comps' => null,
-                    'details_url' => 'https://stock.adobe.com/280369835?as_channel=affiliate&as_source=api&as_content=7f387d439e4f4e64ae42060fc571c456',
-                    'template_type_id' => null,
-                    'template_category_ids' => null,
-                    'marketing_text' => null,
-                    'description' => null,
-                    'size_bytes' => null,
-                    'premium_level_id' => 0,
-                    'is_premium' => null,
-                    'licenses' => null,
-                    'video_preview_url' => null,
-                    'VideoPreviewHeight' => null,
-                    'video_preview_width' => null,
-                    'video_preview_content_length' => null,
-                    'video_preview_content_type' => null,
-                    'video_small_preview_url' => null,
-                    'video_small_preview_height' => null,
-                    'video_small_preview_width' => null,
-                    'video_small_preview_content_length' => null,
-                    'video_small_preview_content_type' => null,
-                ],
-                21 => [
-                    'id' => 270406788,
-                    'title' => 'Still life with fresh medlar on the metallic plate',
-                    'creator_id' => 204226399,
-                    'creator_name' => 'zakiroff',
-                    'creation_date' => '2019-05-29 15:19:36',
-                    'country_name' => 'Kazakhstan',
-                    'thumbnail_url' => null,
-                    'thumbnail_html_tag' => null,
-                    'thumbnail_width' => null,
-                    'thumbnail_height' => null,
-                    'thumbnail_110_url' => null,
-                    'thumbnail_110_width' => null,
-                    'thumbnail_110_height' => null,
-                    'thumbnail_160_url' => null,
-                    'thumbnail_160_width' => null,
-                    'thumbnail_160_height' => null,
-                    'thumbnail_220_url' => null,
-                    'thumbnail_220_width' => null,
-                    'thumbnail_220_height' => null,
-                    'thumbnail_240_url' => 'https://t4.ftcdn.net/jpg/02/70/40/67/240_F_270406788_o7OBZ9o3URMqWSOp58XqI2YOKINnr2HV.jpg',
-                    'thumbnail_240_width' => null,
-                    'thumbnail_240_height' => null,
-                    'thumbnail_500_url' => 'https://as2.ftcdn.net/jpg/02/70/40/67/500_F_270406788_o7OBZ9o3URMqWSOp58XqI2YOKINnr2HV.jpg',
-                    'thumbnail_500_width' => null,
-                    'thumbnail_500_height' => null,
-                    'thumbnail_1000_url' => null,
-                    'thumbnail_1000_width' => null,
-                    'thumbnail_1000_height' => null,
-                    'media_type_id' => 1,
-                    'width' => 3840,
-                    'height' => 5265,
-                    'is_licensed' => null,
-                    'comp_url' => 'https://stock.adobe.com/Rest/Libraries/Watermarked/Download/270406788/2',
-                    'comp_width' => null,
-                    'comp_height' => null,
-                    'nb_views' => null,
-                    'nb_downloads' => null,
-                    'category' => [
-                        'id' => 289,
-                        'name' => 'Food',
-                        'link' => null,
-                    ],
-                    'keywords' => [
-                        0 => [
-                            'name' => 'medlar',
-                        ],
-                        1 => [
-                            'name' => 'fruit',
-                        ],
-                        2 => [
-                            'name' => 'still life',
-                        ],
-                        3 => [
-                            'name' => 'table',
-                        ],
-                        4 => [
-                            'name' => 'yellow',
-                        ],
-                        5 => [
-                            'name' => 'blue',
-                        ],
-                        6 => [
-                            'name' => 'plate',
-                        ],
-                        7 => [
-                            'name' => 'whole',
-                        ],
-                        8 => [
-                            'name' => 'halved',
-                        ],
-                        9 => [
-                            'name' => 'tablecloth',
-                        ],
-                        10 => [
-                            'name' => 'textile',
-                        ],
-                        11 => [
-                            'name' => 'dark',
-                        ],
-                        12 => [
-                            'name' => 'art',
-                        ],
-                        13 => [
-                            'name' => 'arrangement',
-                        ],
-                        14 => [
-                            'name' => 'food',
-                        ],
-                        15 => [
-                            'name' => 'indoor',
-                        ],
-                        16 => [
-                            'name' => 'day',
-                        ],
-                        17 => [
-                            'name' => 'no people',
-                        ],
-                        18 => [
-                            'name' => 'nobody',
-                        ],
-                    ],
-                    'has_releases' => null,
-                    'asset_type_id' => null,
-                    'vector_type' => null,
-                    'content_type' => 'image/jpeg',
-                    'framerate' => null,
-                    'duration' => null,
-                    'stock_id' => null,
-                    'comps' => null,
-                    'details_url' => 'https://stock.adobe.com/270406788?as_channel=affiliate&as_source=api&as_content=7f387d439e4f4e64ae42060fc571c456',
-                    'template_type_id' => null,
-                    'template_category_ids' => null,
-                    'marketing_text' => null,
-                    'description' => null,
-                    'size_bytes' => null,
-                    'premium_level_id' => 0,
-                    'is_premium' => null,
-                    'licenses' => null,
-                    'video_preview_url' => null,
-                    'VideoPreviewHeight' => null,
-                    'video_preview_width' => null,
-                    'video_preview_content_length' => null,
-                    'video_preview_content_type' => null,
-                    'video_small_preview_url' => null,
-                    'video_small_preview_height' => null,
-                    'video_small_preview_width' => null,
-                    'video_small_preview_content_length' => null,
-                    'video_small_preview_content_type' => null,
-                ],
-                22 => [
-                    'id' => 262463286,
-                    'title' => 'Two boys splashing in water near the Damnoen Saduak Floating Market',
-                    'creator_id' => 203814079,
-                    'creator_name' => 'PRASERT',
-                    'creation_date' => '2019-04-17 11:18:30',
-                    'country_name' => 'Thailand',
-                    'thumbnail_url' => null,
-                    'thumbnail_html_tag' => null,
-                    'thumbnail_width' => null,
-                    'thumbnail_height' => null,
-                    'thumbnail_110_url' => null,
-                    'thumbnail_110_width' => null,
-                    'thumbnail_110_height' => null,
-                    'thumbnail_160_url' => null,
-                    'thumbnail_160_width' => null,
-                    'thumbnail_160_height' => null,
-                    'thumbnail_220_url' => null,
-                    'thumbnail_220_width' => null,
-                    'thumbnail_220_height' => null,
-                    'thumbnail_240_url' => 'https://t3.ftcdn.net/jpg/02/62/46/32/240_F_262463286_oHnZM7z3NjoLxzZI8Z2tvx8kFjgPxGTc.jpg',
-                    'thumbnail_240_width' => null,
-                    'thumbnail_240_height' => null,
-                    'thumbnail_500_url' => 'https://as1.ftcdn.net/jpg/02/62/46/32/500_F_262463286_oHnZM7z3NjoLxzZI8Z2tvx8kFjgPxGTc.jpg',
-                    'thumbnail_500_width' => null,
-                    'thumbnail_500_height' => null,
-                    'thumbnail_1000_url' => null,
-                    'thumbnail_1000_width' => null,
-                    'thumbnail_1000_height' => null,
-                    'media_type_id' => 1,
-                    'width' => 5472,
-                    'height' => 3648,
-                    'is_licensed' => null,
-                    'comp_url' => 'https://stock.adobe.com/Rest/Libraries/Watermarked/Download/262463286/2',
-                    'comp_width' => null,
-                    'comp_height' => null,
-                    'nb_views' => null,
-                    'nb_downloads' => null,
-                    'category' => [
-                        'id' => 498,
-                        'name' => 'Hobbies and Leisure',
-                        'link' => null,
-                    ],
-                    'keywords' => [
-                        0 => [
-                            'name' => 'boy',
-                        ],
-                        1 => [
-                            'name' => 'children',
-                        ],
-                        2 => [
-                            'name' => 'elementary',
-                        ],
-                        3 => [
-                            'name' => 'thai',
-                        ],
-                        4 => [
-                            'name' => 'pair',
-                        ],
-                        5 => [
-                            'name' => 'water',
-                        ],
-                        6 => [
-                            'name' => 'splash',
-                        ],
-                        7 => [
-                            'name' => 'play',
-                        ],
-                        8 => [
-                            'name' => 'happiness',
-                        ],
-                        9 => [
-                            'name' => 'smile',
-                        ],
-                        10 => [
-                            'name' => 'shirtless',
-                        ],
-                        11 => [
-                            'name' => 'to sit',
-                        ],
-                        12 => [
-                            'name' => 'sun',
-                        ],
-                        13 => [
-                            'name' => 'sunlight',
-                        ],
-                        14 => [
-                            'name' => 'sunset',
-                        ],
-                        15 => [
-                            'name' => 'carefree',
-                        ],
-                        16 => [
-                            'name' => 'friendship',
-                        ],
-                        17 => [
-                            'name' => 'fun',
-                        ],
-                        18 => [
-                            'name' => 'canal',
-                        ],
-                        19 => [
-                            'name' => 'siam',
-                        ],
-                        20 => [
-                            'name' => 'outdoors',
-                        ],
-                        21 => [
-                            'name' => '2',
-                        ],
-                        22 => [
-                            'name' => 'people',
-                        ],
-                    ],
-                    'has_releases' => null,
-                    'asset_type_id' => null,
-                    'vector_type' => null,
-                    'content_type' => 'image/jpeg',
-                    'framerate' => null,
-                    'duration' => null,
-                    'stock_id' => null,
-                    'comps' => null,
-                    'details_url' => 'https://stock.adobe.com/262463286?as_channel=affiliate&as_source=api&as_content=7f387d439e4f4e64ae42060fc571c456',
-                    'template_type_id' => null,
-                    'template_category_ids' => null,
-                    'marketing_text' => null,
-                    'description' => null,
-                    'size_bytes' => null,
-                    'premium_level_id' => 0,
-                    'is_premium' => null,
-                    'licenses' => null,
-                    'video_preview_url' => null,
-                    'VideoPreviewHeight' => null,
-                    'video_preview_width' => null,
-                    'video_preview_content_length' => null,
-                    'video_preview_content_type' => null,
-                    'video_small_preview_url' => null,
-                    'video_small_preview_height' => null,
-                    'video_small_preview_width' => null,
-                    'video_small_preview_content_length' => null,
-                    'video_small_preview_content_type' => null,
-                ],
-                23 => [
-                    'id' => 299767794,
-                    'title' => 'Close up of woman smoking',
-                    'creator_id' => 204298082,
-                    'creator_name' => 'oneinchpunch',
-                    'creation_date' => '2019-11-01 15:07:07',
-                    'country_name' => 'Italy',
-                    'thumbnail_url' => null,
-                    'thumbnail_html_tag' => null,
-                    'thumbnail_width' => null,
-                    'thumbnail_height' => null,
-                    'thumbnail_110_url' => null,
-                    'thumbnail_110_width' => null,
-                    'thumbnail_110_height' => null,
-                    'thumbnail_160_url' => null,
-                    'thumbnail_160_width' => null,
-                    'thumbnail_160_height' => null,
-                    'thumbnail_220_url' => null,
-                    'thumbnail_220_width' => null,
-                    'thumbnail_220_height' => null,
-                    'thumbnail_240_url' => 'https://t4.ftcdn.net/jpg/02/99/76/77/240_F_299767794_MFtLlxTxLJjPCegl4qxUhkfq28SzgzJt.jpg',
-                    'thumbnail_240_width' => null,
-                    'thumbnail_240_height' => null,
-                    'thumbnail_500_url' => 'https://as2.ftcdn.net/jpg/02/99/76/77/500_F_299767794_MFtLlxTxLJjPCegl4qxUhkfq28SzgzJt.jpg',
-                    'thumbnail_500_width' => null,
-                    'thumbnail_500_height' => null,
-                    'thumbnail_1000_url' => null,
-                    'thumbnail_1000_width' => null,
-                    'thumbnail_1000_height' => null,
-                    'media_type_id' => 1,
-                    'width' => 4945,
-                    'height' => 3234,
-                    'is_licensed' => null,
-                    'comp_url' => 'https://stock.adobe.com/Rest/Libraries/Watermarked/Download/299767794/2',
-                    'comp_width' => null,
-                    'comp_height' => null,
-                    'nb_views' => null,
-                    'nb_downloads' => null,
-                    'category' => [
-                        'id' => 695,
-                        'name' => 'People',
-                        'link' => null,
-                    ],
-                    'keywords' => [
-                        0 => [
-                            'name' => 'woman',
-                        ],
-                        1 => [
-                            'name' => 'adult',
-                        ],
-                        2 => [
-                            'name' => 'young',
-                        ],
-                        3 => [
-                            'name' => 'latin',
-                        ],
-                        4 => [
-                            'name' => 'alone',
-                        ],
-                        5 => [
-                            'name' => 'smoke',
-                        ],
-                        6 => [
-                            'name' => 'lip',
-                        ],
-                        7 => [
-                            'name' => 'mouth',
-                        ],
-                        8 => [
-                            'name' => 'lipstick',
-                        ],
-                        9 => [
-                            'name' => 'red',
-                        ],
-                        10 => [
-                            'name' => 'freckle',
-                        ],
-                        11 => [
-                            'name' => 'fur',
-                        ],
-                        12 => [
-                            'name' => 'unhealthy',
-                        ],
-                        13 => [
-                            'name' => 'lifestyle',
-                        ],
-                        14 => [
-                            'name' => 'addiction',
-                        ],
-                        15 => [
-                            'name' => 'party',
-                        ],
-                        16 => [
-                            'name' => 'closeup',
-                        ],
-                        17 => [
-                            'name' => 'incomplete',
-                        ],
-                        18 => [
-                            'name' => 'indoor',
-                        ],
-                        19 => [
-                            'name' => 'day',
-                        ],
-                        20 => [
-                            'name' => '1',
-                        ],
-                        21 => [
-                            'name' => 'people',
-                        ],
-                    ],
-                    'has_releases' => null,
-                    'asset_type_id' => null,
-                    'vector_type' => null,
-                    'content_type' => 'image/jpeg',
-                    'framerate' => null,
-                    'duration' => null,
-                    'stock_id' => null,
-                    'comps' => null,
-                    'details_url' => 'https://stock.adobe.com/299767794?as_channel=affiliate&as_source=api&as_content=7f387d439e4f4e64ae42060fc571c456',
-                    'template_type_id' => null,
-                    'template_category_ids' => null,
-                    'marketing_text' => null,
-                    'description' => null,
-                    'size_bytes' => null,
-                    'premium_level_id' => 0,
-                    'is_premium' => null,
-                    'licenses' => null,
-                    'video_preview_url' => null,
-                    'VideoPreviewHeight' => null,
-                    'video_preview_width' => null,
-                    'video_preview_content_length' => null,
-                    'video_preview_content_type' => null,
-                    'video_small_preview_url' => null,
-                    'video_small_preview_height' => null,
-                    'video_small_preview_width' => null,
-                    'video_small_preview_content_length' => null,
-                    'video_small_preview_content_type' => null,
-                ],
-                24 => [
-                    'id' => 248573004,
-                    'title' => 'Smiling demonstrator with megaphone',
-                    'creator_id' => 204567087,
-                    'creator_name' => 'Rawpixel.com',
-                    'creation_date' => '2019-02-11 06:02:05',
-                    'country_name' => 'United Kingdom of Great Britain and Northern Ireland',
-                    'thumbnail_url' => null,
-                    'thumbnail_html_tag' => null,
-                    'thumbnail_width' => null,
-                    'thumbnail_height' => null,
-                    'thumbnail_110_url' => null,
-                    'thumbnail_110_width' => null,
-                    'thumbnail_110_height' => null,
-                    'thumbnail_160_url' => null,
-                    'thumbnail_160_width' => null,
-                    'thumbnail_160_height' => null,
-                    'thumbnail_220_url' => null,
-                    'thumbnail_220_width' => null,
-                    'thumbnail_220_height' => null,
-                    'thumbnail_240_url' => 'https://t3.ftcdn.net/jpg/02/48/57/30/240_F_248573004_gH8C08zbZekY4ZSGSeOdtvf2QbY9oFQK.jpg',
-                    'thumbnail_240_width' => null,
-                    'thumbnail_240_height' => null,
-                    'thumbnail_500_url' => 'https://as1.ftcdn.net/jpg/02/48/57/30/500_F_248573004_gH8C08zbZekY4ZSGSeOdtvf2QbY9oFQK.jpg',
-                    'thumbnail_500_width' => null,
-                    'thumbnail_500_height' => null,
-                    'thumbnail_1000_url' => null,
-                    'thumbnail_1000_width' => null,
-                    'thumbnail_1000_height' => null,
-                    'media_type_id' => 1,
-                    'width' => 6542,
-                    'height' => 4361,
-                    'is_licensed' => null,
-                    'comp_url' => 'https://stock.adobe.com/Rest/Libraries/Watermarked/Download/248573004/2',
-                    'comp_width' => null,
-                    'comp_height' => null,
-                    'nb_views' => null,
-                    'nb_downloads' => null,
-                    'category' => [
-                        'id' => 695,
-                        'name' => 'People',
-                        'link' => null,
-                    ],
-                    'keywords' => [
-                        0 => [
-                            'name' => 'woman',
-                        ],
-                        1 => [
-                            'name' => 'adult',
-                        ],
-                        2 => [
-                            'name' => 'young',
-                        ],
-                        3 => [
-                            'name' => 'african american',
-                        ],
-                        4 => [
-                            'name' => 'activist',
-                        ],
-                        5 => [
-                            'name' => 'megaphone',
-                        ],
-                        6 => [
-                            'name' => 'protest',
-                        ],
-                        7 => [
-                            'name' => 'eyeglass',
-                        ],
-                        8 => [
-                            'name' => 'headband',
-                        ],
-                        9 => [
-                            'name' => 'smile',
-                        ],
-                        10 => [
-                            'name' => 'cheerful',
-                        ],
-                        11 => [
-                            'name' => 'peace',
-                        ],
-                        12 => [
-                            'name' => 'community',
-                        ],
-                        13 => [
-                            'name' => 'social issue',
-                        ],
-                        14 => [
-                            'name' => 'lifestyle',
-                        ],
-                        15 => [
-                            'name' => 'equality',
-                        ],
-                        16 => [
-                            'name' => 'leadership',
-                        ],
-                        17 => [
-                            'name' => 'strike',
-                        ],
-                        18 => [
-                            'name' => 'kansas city',
-                        ],
-                        19 => [
-                            'name' => 'kansas',
-                        ],
-                        20 => [
-                            'name' => 'us',
-                        ],
-                        21 => [
-                            'name' => 'us',
-                        ],
-                        22 => [
-                            'name' => 'outdoors',
-                        ],
-                        23 => [
-                            'name' => 'day',
-                        ],
-                        24 => [
-                            'name' => 'group',
-                        ],
-                        25 => [
-                            'name' => '4',
-                        ],
-                        26 => [
-                            'name' => 'people',
-                        ],
-                    ],
-                    'has_releases' => null,
-                    'asset_type_id' => null,
-                    'vector_type' => null,
-                    'content_type' => 'image/jpeg',
-                    'framerate' => null,
-                    'duration' => null,
-                    'stock_id' => null,
-                    'comps' => null,
-                    'details_url' => 'https://stock.adobe.com/248573004?as_channel=affiliate&as_source=api&as_content=7f387d439e4f4e64ae42060fc571c456',
-                    'template_type_id' => null,
-                    'template_category_ids' => null,
-                    'marketing_text' => null,
-                    'description' => null,
-                    'size_bytes' => null,
-                    'premium_level_id' => 0,
-                    'is_premium' => null,
-                    'licenses' => null,
-                    'video_preview_url' => null,
-                    'VideoPreviewHeight' => null,
-                    'video_preview_width' => null,
-                    'video_preview_content_length' => null,
-                    'video_preview_content_type' => null,
-                    'video_small_preview_url' => null,
-                    'video_small_preview_height' => null,
-                    'video_small_preview_width' => null,
-                    'video_small_preview_content_length' => null,
-                    'video_small_preview_content_type' => null,
-                ],
-                25 => [
-                    'id' => 284151546,
-                    'title' => 'Stylish woman wearing sunglasses drinking juice',
-                    'creator_id' => 206795825,
-                    'creator_name' => 'PhotoBook',
-                    'creation_date' => '2019-08-15 17:02:20',
-                    'country_name' => 'Ukraine',
-                    'thumbnail_url' => null,
-                    'thumbnail_html_tag' => null,
-                    'thumbnail_width' => null,
-                    'thumbnail_height' => null,
-                    'thumbnail_110_url' => null,
-                    'thumbnail_110_width' => null,
-                    'thumbnail_110_height' => null,
-                    'thumbnail_160_url' => null,
-                    'thumbnail_160_width' => null,
-                    'thumbnail_160_height' => null,
-                    'thumbnail_220_url' => null,
-                    'thumbnail_220_width' => null,
-                    'thumbnail_220_height' => null,
-                    'thumbnail_240_url' => 'https://t4.ftcdn.net/jpg/02/84/15/15/240_F_284151546_ORK9qa0ACvEXMwE9qN5BymHxpFYj7tqM.jpg',
-                    'thumbnail_240_width' => null,
-                    'thumbnail_240_height' => null,
-                    'thumbnail_500_url' => 'https://as2.ftcdn.net/jpg/02/84/15/15/500_F_284151546_ORK9qa0ACvEXMwE9qN5BymHxpFYj7tqM.jpg',
-                    'thumbnail_500_width' => null,
-                    'thumbnail_500_height' => null,
-                    'thumbnail_1000_url' => null,
-                    'thumbnail_1000_width' => null,
-                    'thumbnail_1000_height' => null,
-                    'media_type_id' => 1,
-                    'width' => 3057,
-                    'height' => 4585,
-                    'is_licensed' => null,
-                    'comp_url' => 'https://stock.adobe.com/Rest/Libraries/Watermarked/Download/284151546/2',
-                    'comp_width' => null,
-                    'comp_height' => null,
-                    'nb_views' => null,
-                    'nb_downloads' => null,
-                    'category' => [
-                        'id' => 695,
-                        'name' => 'People',
-                        'link' => null,
-                    ],
-                    'keywords' => [
-                        0 => [
-                            'name' => 'woman',
-                        ],
-                        1 => [
-                            'name' => 'adult',
-                        ],
-                        2 => [
-                            'name' => 'young',
-                        ],
-                        3 => [
-                            'name' => 'alone',
-                        ],
-                        4 => [
-                            'name' => 'smile',
-                        ],
-                        5 => [
-                            'name' => 'sunglasses',
-                        ],
-                        6 => [
-                            'name' => 'top',
-                        ],
-                        7 => [
-                            'name' => 'bottle',
-                        ],
-                        8 => [
-                            'name' => 'straw',
-                        ],
-                        9 => [
-                            'name' => 'beverage',
-                        ],
-                        10 => [
-                            'name' => 'juice',
-                        ],
-                        11 => [
-                            'name' => 'summer',
-                        ],
-                        12 => [
-                            'name' => 'relaxation',
-                        ],
-                        13 => [
-                            'name' => 'stylish',
-                        ],
-                        14 => [
-                            'name' => 'fashion',
-                        ],
-                        15 => [
-                            'name' => 'fashionable',
-                        ],
-                        16 => [
-                            'name' => 'cheerful',
-                        ],
-                        17 => [
-                            'name' => 'lifestyle',
-                        ],
-                        18 => [
-                            'name' => 'outdoors',
-                        ],
-                        19 => [
-                            'name' => 'day',
-                        ],
-                        20 => [
-                            'name' => '1',
-                        ],
-                        21 => [
-                            'name' => 'people',
-                        ],
-                    ],
-                    'has_releases' => null,
-                    'asset_type_id' => null,
-                    'vector_type' => null,
-                    'content_type' => 'image/jpeg',
-                    'framerate' => null,
-                    'duration' => null,
-                    'stock_id' => null,
-                    'comps' => null,
-                    'details_url' => 'https://stock.adobe.com/284151546?as_channel=affiliate&as_source=api&as_content=7f387d439e4f4e64ae42060fc571c456',
-                    'template_type_id' => null,
-                    'template_category_ids' => null,
-                    'marketing_text' => null,
-                    'description' => null,
-                    'size_bytes' => null,
-                    'premium_level_id' => 0,
-                    'is_premium' => null,
-                    'licenses' => null,
-                    'video_preview_url' => null,
-                    'VideoPreviewHeight' => null,
-                    'video_preview_width' => null,
-                    'video_preview_content_length' => null,
-                    'video_preview_content_type' => null,
-                    'video_small_preview_url' => null,
-                    'video_small_preview_height' => null,
-                    'video_small_preview_width' => null,
-                    'video_small_preview_content_length' => null,
-                    'video_small_preview_content_type' => null,
-                ],
-                26 => [
-                    'id' => 276148279,
-                    'title' => 'Close up of broken egg in carton',
-                    'creator_id' => 206713618,
-                    'creator_name' => 'LIGHTFIELD STUDIOS',
-                    'creation_date' => '2019-07-01 15:38:24',
-                    'country_name' => 'United States of America',
-                    'thumbnail_url' => null,
-                    'thumbnail_html_tag' => null,
-                    'thumbnail_width' => null,
-                    'thumbnail_height' => null,
-                    'thumbnail_110_url' => null,
-                    'thumbnail_110_width' => null,
-                    'thumbnail_110_height' => null,
-                    'thumbnail_160_url' => null,
-                    'thumbnail_160_width' => null,
-                    'thumbnail_160_height' => null,
-                    'thumbnail_220_url' => null,
-                    'thumbnail_220_width' => null,
-                    'thumbnail_220_height' => null,
-                    'thumbnail_240_url' => 'https://t3.ftcdn.net/jpg/02/76/14/82/240_F_276148279_2WgA5xFgfZibdGSpELjpXkQzEmrjKk4D.jpg',
-                    'thumbnail_240_width' => null,
-                    'thumbnail_240_height' => null,
-                    'thumbnail_500_url' => 'https://as1.ftcdn.net/jpg/02/76/14/82/500_F_276148279_2WgA5xFgfZibdGSpELjpXkQzEmrjKk4D.jpg',
-                    'thumbnail_500_width' => null,
-                    'thumbnail_500_height' => null,
-                    'thumbnail_1000_url' => null,
-                    'thumbnail_1000_width' => null,
-                    'thumbnail_1000_height' => null,
-                    'media_type_id' => 1,
-                    'width' => 4480,
-                    'height' => 6720,
-                    'is_licensed' => null,
-                    'comp_url' => 'https://stock.adobe.com/Rest/Libraries/Watermarked/Download/276148279/2',
-                    'comp_width' => null,
-                    'comp_height' => null,
-                    'nb_views' => null,
-                    'nb_downloads' => null,
-                    'category' => [
-                        'id' => 289,
-                        'name' => 'Food',
-                        'link' => null,
-                    ],
-                    'keywords' => [
-                        0 => [
-                            'name' => 'egg',
-                        ],
-                        1 => [
-                            'name' => 'yolk',
-                        ],
-                        2 => [
-                            'name' => 'broken',
-                        ],
-                        3 => [
-                            'name' => 'whole',
-                        ],
-                        4 => [
-                            'name' => 'cardbox',
-                        ],
-                        5 => [
-                            'name' => 'separation',
-                        ],
-                        6 => [
-                            'name' => 'individuality',
-                        ],
-                        7 => [
-                            'name' => 'shell',
-                        ],
-                        8 => [
-                            'name' => 'fragility',
-                        ],
-                        9 => [
-                            'name' => 'food',
-                        ],
-                        10 => [
-                            'name' => 'closeup',
-                        ],
-                        11 => [
-                            'name' => 'indoor',
-                        ],
-                        12 => [
-                            'name' => 'day',
-                        ],
-                        13 => [
-                            'name' => 'no people',
-                        ],
-                        14 => [
-                            'name' => 'nobody',
-                        ],
-                    ],
-                    'has_releases' => null,
-                    'asset_type_id' => null,
-                    'vector_type' => null,
-                    'content_type' => 'image/jpeg',
-                    'framerate' => null,
-                    'duration' => null,
-                    'stock_id' => null,
-                    'comps' => null,
-                    'details_url' => 'https://stock.adobe.com/276148279?as_channel=affiliate&as_source=api&as_content=7f387d439e4f4e64ae42060fc571c456',
-                    'template_type_id' => null,
-                    'template_category_ids' => null,
-                    'marketing_text' => null,
-                    'description' => null,
-                    'size_bytes' => null,
-                    'premium_level_id' => 0,
-                    'is_premium' => null,
-                    'licenses' => null,
-                    'video_preview_url' => null,
-                    'VideoPreviewHeight' => null,
-                    'video_preview_width' => null,
-                    'video_preview_content_length' => null,
-                    'video_preview_content_type' => null,
-                    'video_small_preview_url' => null,
-                    'video_small_preview_height' => null,
-                    'video_small_preview_width' => null,
-                    'video_small_preview_content_length' => null,
-                    'video_small_preview_content_type' => null,
-                ],
-                27 => [
-                    'id' => 282277048,
-                    'title' => 'Young woman running in New York, United States',
-                    'creator_id' => 204298082,
-                    'creator_name' => 'oneinchpunch',
-                    'creation_date' => '2019-08-05 17:01:51',
-                    'country_name' => 'Italy',
-                    'thumbnail_url' => null,
-                    'thumbnail_html_tag' => null,
-                    'thumbnail_width' => null,
-                    'thumbnail_height' => null,
-                    'thumbnail_110_url' => null,
-                    'thumbnail_110_width' => null,
-                    'thumbnail_110_height' => null,
-                    'thumbnail_160_url' => null,
-                    'thumbnail_160_width' => null,
-                    'thumbnail_160_height' => null,
-                    'thumbnail_220_url' => null,
-                    'thumbnail_220_width' => null,
-                    'thumbnail_220_height' => null,
-                    'thumbnail_240_url' => 'https://t3.ftcdn.net/jpg/02/82/27/70/240_F_282277048_mKzq2hT5L1DtHBRmT3JkoNcJ9Y7a0wad.jpg',
-                    'thumbnail_240_width' => null,
-                    'thumbnail_240_height' => null,
-                    'thumbnail_500_url' => 'https://as1.ftcdn.net/jpg/02/82/27/70/500_F_282277048_mKzq2hT5L1DtHBRmT3JkoNcJ9Y7a0wad.jpg',
-                    'thumbnail_500_width' => null,
-                    'thumbnail_500_height' => null,
-                    'thumbnail_1000_url' => null,
-                    'thumbnail_1000_width' => null,
-                    'thumbnail_1000_height' => null,
-                    'media_type_id' => 1,
-                    'width' => 5891,
-                    'height' => 3933,
-                    'is_licensed' => null,
-                    'comp_url' => 'https://stock.adobe.com/Rest/Libraries/Watermarked/Download/282277048/2',
-                    'comp_width' => null,
-                    'comp_height' => null,
-                    'nb_views' => null,
-                    'nb_downloads' => null,
-                    'category' => [
-                        'id' => 922,
-                        'name' => 'Sports',
-                        'link' => null,
-                    ],
-                    'keywords' => [
-                        0 => [
-                            'name' => 'young',
-                        ],
-                        1 => [
-                            'name' => 'woman',
-                        ],
-                        2 => [
-                            'name' => 'black',
-                        ],
-                        3 => [
-                            'name' => 'african american',
-                        ],
-                        4 => [
-                            'name' => 'running',
-                        ],
-                        5 => [
-                            'name' => 'sport',
-                        ],
-                        6 => [
-                            'name' => 'jog',
-                        ],
-                        7 => [
-                            'name' => 'jogger',
-                        ],
-                        8 => [
-                            'name' => 'jogging',
-                        ],
-                        9 => [
-                            'name' => 'fit',
-                        ],
-                        10 => [
-                            'name' => 'fitness',
-                        ],
-                        11 => [
-                            'name' => 'training',
-                        ],
-                        12 => [
-                            'name' => 'activewear',
-                        ],
-                        13 => [
-                            'name' => 'brick',
-                        ],
-                        14 => [
-                            'name' => 'wall',
-                        ],
-                        15 => [
-                            'name' => 'speed',
-                        ],
-                        16 => [
-                            'name' => 'outdoors',
-                        ],
-                        17 => [
-                            'name' => 'day',
-                        ],
-                        18 => [
-                            'name' => 'people',
-                        ],
-                        19 => [
-                            'name' => '1',
-                        ],
-                        20 => [
-                            'name' => 'lifestyle',
-                        ],
-                        21 => [
-                            'name' => 'new york city',
-                        ],
-                        22 => [
-                            'name' => 'us',
-                        ],
-                        23 => [
-                            'name' => 'us',
-                        ],
-                        24 => [
-                            'name' => 'alone',
-                        ],
-                    ],
-                    'has_releases' => null,
-                    'asset_type_id' => null,
-                    'vector_type' => null,
-                    'content_type' => 'image/jpeg',
-                    'framerate' => null,
-                    'duration' => null,
-                    'stock_id' => null,
-                    'comps' => null,
-                    'details_url' => 'https://stock.adobe.com/282277048?as_channel=affiliate&as_source=api&as_content=7f387d439e4f4e64ae42060fc571c456',
-                    'template_type_id' => null,
-                    'template_category_ids' => null,
-                    'marketing_text' => null,
-                    'description' => null,
-                    'size_bytes' => null,
-                    'premium_level_id' => 0,
-                    'is_premium' => null,
-                    'licenses' => null,
-                    'video_preview_url' => null,
-                    'VideoPreviewHeight' => null,
-                    'video_preview_width' => null,
-                    'video_preview_content_length' => null,
-                    'video_preview_content_type' => null,
-                    'video_small_preview_url' => null,
-                    'video_small_preview_height' => null,
-                    'video_small_preview_width' => null,
-                    'video_small_preview_content_length' => null,
-                    'video_small_preview_content_type' => null,
-                ],
-                28 => [
-                    'id' => 268514505,
-                    'title' => 'Close up of Gelada monkey, Ethiopia',
-                    'creator_id' => 204589027,
-                    'creator_name' => 'giedriius',
-                    'creation_date' => '2019-05-18 15:40:27',
-                    'country_name' => 'United Kingdom of Great Britain and Northern Ireland',
-                    'thumbnail_url' => null,
-                    'thumbnail_html_tag' => null,
-                    'thumbnail_width' => null,
-                    'thumbnail_height' => null,
-                    'thumbnail_110_url' => null,
-                    'thumbnail_110_width' => null,
-                    'thumbnail_110_height' => null,
-                    'thumbnail_160_url' => null,
-                    'thumbnail_160_width' => null,
-                    'thumbnail_160_height' => null,
-                    'thumbnail_220_url' => null,
-                    'thumbnail_220_width' => null,
-                    'thumbnail_220_height' => null,
-                    'thumbnail_240_url' => 'https://t4.ftcdn.net/jpg/02/68/51/45/240_F_268514505_GUD9o8SdBo3cAgmKkvBsMyZZYRweJ7LP.jpg',
-                    'thumbnail_240_width' => null,
-                    'thumbnail_240_height' => null,
-                    'thumbnail_500_url' => 'https://as2.ftcdn.net/jpg/02/68/51/45/500_F_268514505_GUD9o8SdBo3cAgmKkvBsMyZZYRweJ7LP.jpg',
-                    'thumbnail_500_width' => null,
-                    'thumbnail_500_height' => null,
-                    'thumbnail_1000_url' => null,
-                    'thumbnail_1000_width' => null,
-                    'thumbnail_1000_height' => null,
-                    'media_type_id' => 1,
-                    'width' => 2600,
-                    'height' => 3348,
-                    'is_licensed' => null,
-                    'comp_url' => 'https://stock.adobe.com/Rest/Libraries/Watermarked/Download/268514505/2',
-                    'comp_width' => null,
-                    'comp_height' => null,
-                    'nb_views' => null,
-                    'nb_downloads' => null,
-                    'category' => [
-                        'id' => 1,
-                        'name' => 'Animals',
-                        'link' => null,
-                    ],
-                    'keywords' => [
-                        0 => [
-                            'name' => 'monkey',
-                        ],
-                        1 => [
-                            'name' => 'endemic',
-                        ],
-                        2 => [
-                            'name' => 'ethiopia',
-                        ],
-                        3 => [
-                            'name' => 'fur',
-                        ],
-                        4 => [
-                            'name' => 'nature',
-                        ],
-                        5 => [
-                            'name' => 'wildlife',
-                        ],
-                        6 => [
-                            'name' => 'face',
-                        ],
-                        7 => [
-                            'name' => 'hair',
-                        ],
-                        8 => [
-                            'name' => 'brown',
-                        ],
-                        9 => [
-                            'name' => 'portrait',
-                        ],
-                        10 => [
-                            'name' => 'closeup',
-                        ],
-                        11 => [
-                            'name' => 'outdoors',
-                        ],
-                        12 => [
-                            'name' => 'day',
-                        ],
-                        13 => [
-                            'name' => 'animal',
-                        ],
-                        14 => [
-                            'name' => '1',
-                        ],
-                        15 => [
-                            'name' => 'no people',
-                        ],
-                        16 => [
-                            'name' => 'nobody',
-                        ],
-                    ],
-                    'has_releases' => null,
-                    'asset_type_id' => null,
-                    'vector_type' => null,
-                    'content_type' => 'image/jpeg',
-                    'framerate' => null,
-                    'duration' => null,
-                    'stock_id' => null,
-                    'comps' => null,
-                    'details_url' => 'https://stock.adobe.com/268514505?as_channel=affiliate&as_source=api&as_content=7f387d439e4f4e64ae42060fc571c456',
-                    'template_type_id' => null,
-                    'template_category_ids' => null,
-                    'marketing_text' => null,
-                    'description' => null,
-                    'size_bytes' => null,
-                    'premium_level_id' => 0,
-                    'is_premium' => null,
-                    'licenses' => null,
-                    'video_preview_url' => null,
-                    'VideoPreviewHeight' => null,
-                    'video_preview_width' => null,
-                    'video_preview_content_length' => null,
-                    'video_preview_content_type' => null,
-                    'video_small_preview_url' => null,
-                    'video_small_preview_height' => null,
-                    'video_small_preview_width' => null,
-                    'video_small_preview_content_length' => null,
-                    'video_small_preview_content_type' => null,
-                ],
-                29 => [
-                    'id' => 242506295,
-                    'title' => 'Woman wearing Muslimah sportswear stretching outdoors',
-                    'creator_id' => 224608,
-                    'creator_name' => 'Jacob Lund ',
-                    'creation_date' => '2019-01-09 17:29:38',
-                    'country_name' => 'Denmark',
-                    'thumbnail_url' => null,
-                    'thumbnail_html_tag' => null,
-                    'thumbnail_width' => null,
-                    'thumbnail_height' => null,
-                    'thumbnail_110_url' => null,
-                    'thumbnail_110_width' => null,
-                    'thumbnail_110_height' => null,
-                    'thumbnail_160_url' => null,
-                    'thumbnail_160_width' => null,
-                    'thumbnail_160_height' => null,
-                    'thumbnail_220_url' => null,
-                    'thumbnail_220_width' => null,
-                    'thumbnail_220_height' => null,
-                    'thumbnail_240_url' => 'https://t3.ftcdn.net/jpg/02/42/50/62/240_F_242506295_nQERxjjiEfGqXzbwtIZfAdNXQzcDVsm8.jpg',
-                    'thumbnail_240_width' => null,
-                    'thumbnail_240_height' => null,
-                    'thumbnail_500_url' => 'https://as1.ftcdn.net/jpg/02/42/50/62/500_F_242506295_nQERxjjiEfGqXzbwtIZfAdNXQzcDVsm8.jpg',
-                    'thumbnail_500_width' => null,
-                    'thumbnail_500_height' => null,
-                    'thumbnail_1000_url' => null,
-                    'thumbnail_1000_width' => null,
-                    'thumbnail_1000_height' => null,
-                    'media_type_id' => 1,
-                    'width' => 8008,
-                    'height' => 5339,
-                    'is_licensed' => null,
-                    'comp_url' => 'https://stock.adobe.com/Rest/Libraries/Watermarked/Download/242506295/2',
-                    'comp_width' => null,
-                    'comp_height' => null,
-                    'nb_views' => null,
-                    'nb_downloads' => null,
-                    'category' => [
-                        'id' => 922,
-                        'name' => 'Sports',
-                        'link' => null,
-                    ],
-                    'keywords' => [
-                        0 => [
-                            'name' => 'woman',
-                        ],
-                        1 => [
-                            'name' => 'adult',
-                        ],
-                        2 => [
-                            'name' => 'middle eastern',
-                        ],
-                        3 => [
-                            'name' => 'athlete',
-                        ],
-                        4 => [
-                            'name' => 'stretch',
-                        ],
-                        5 => [
-                            'name' => 'exercise',
-                        ],
-                        6 => [
-                            'name' => 'activewear',
-                        ],
-                        7 => [
-                            'name' => 'mobile phone',
-                        ],
-                        8 => [
-                            'name' => 'smartphone',
-                        ],
-                        9 => [
-                            'name' => 'mobile phone',
-                        ],
-                        10 => [
-                            'name' => 'technology',
-                        ],
-                        11 => [
-                            'name' => 'sport',
-                        ],
-                        12 => [
-                            'name' => 'vitality',
-                        ],
-                        13 => [
-                            'name' => 'flexibility',
-                        ],
-                        14 => [
-                            'name' => 'street',
-                        ],
-                        15 => [
-                            'name' => 'outdoors',
-                        ],
-                        16 => [
-                            'name' => 'day',
-                        ],
-                        17 => [
-                            'name' => '1',
-                        ],
-                        18 => [
-                            'name' => 'people',
-                        ],
-                        19 => [
-                            'name' => 'alone',
-                        ],
-                    ],
-                    'has_releases' => null,
-                    'asset_type_id' => null,
-                    'vector_type' => null,
-                    'content_type' => 'image/jpeg',
-                    'framerate' => null,
-                    'duration' => null,
-                    'stock_id' => null,
-                    'comps' => null,
-                    'details_url' => 'https://stock.adobe.com/242506295?as_channel=affiliate&as_source=api&as_content=7f387d439e4f4e64ae42060fc571c456',
-                    'template_type_id' => null,
-                    'template_category_ids' => null,
-                    'marketing_text' => null,
-                    'description' => null,
-                    'size_bytes' => null,
-                    'premium_level_id' => 0,
-                    'is_premium' => null,
-                    'licenses' => null,
-                    'video_preview_url' => null,
-                    'VideoPreviewHeight' => null,
-                    'video_preview_width' => null,
-                    'video_preview_content_length' => null,
-                    'video_preview_content_type' => null,
-                    'video_small_preview_url' => null,
-                    'video_small_preview_height' => null,
-                    'video_small_preview_width' => null,
-                    'video_small_preview_content_length' => null,
-                    'video_small_preview_content_type' => null,
-                ],
-                30 => [
-                    'id' => 299393162,
-                    'title' => 'Yilki horses running in field at sunset, Kayseri, Turkey',
-                    'creator_id' => 200402318,
-                    'creator_name' => 'EvrenKalinbacak',
-                    'creation_date' => '2019-10-30 19:33:03',
-                    'country_name' => 'Turkey',
-                    'thumbnail_url' => null,
-                    'thumbnail_html_tag' => null,
-                    'thumbnail_width' => null,
-                    'thumbnail_height' => null,
-                    'thumbnail_110_url' => null,
-                    'thumbnail_110_width' => null,
-                    'thumbnail_110_height' => null,
-                    'thumbnail_160_url' => null,
-                    'thumbnail_160_width' => null,
-                    'thumbnail_160_height' => null,
-                    'thumbnail_220_url' => null,
-                    'thumbnail_220_width' => null,
-                    'thumbnail_220_height' => null,
-                    'thumbnail_240_url' => 'https://t4.ftcdn.net/jpg/02/99/39/31/240_F_299393162_s4PXhshRvNZj57clBeVMIQz5d68QaA0I.jpg',
-                    'thumbnail_240_width' => null,
-                    'thumbnail_240_height' => null,
-                    'thumbnail_500_url' => 'https://as2.ftcdn.net/jpg/02/99/39/31/500_F_299393162_s4PXhshRvNZj57clBeVMIQz5d68QaA0I.jpg',
-                    'thumbnail_500_width' => null,
-                    'thumbnail_500_height' => null,
-                    'thumbnail_1000_url' => null,
-                    'thumbnail_1000_width' => null,
-                    'thumbnail_1000_height' => null,
-                    'media_type_id' => 1,
-                    'width' => 4410,
-                    'height' => 2940,
-                    'is_licensed' => null,
-                    'comp_url' => 'https://stock.adobe.com/Rest/Libraries/Watermarked/Download/299393162/2',
-                    'comp_width' => null,
-                    'comp_height' => null,
-                    'nb_views' => null,
-                    'nb_downloads' => null,
-                    'category' => [
-                        'id' => 96,
-                        'name' => 'Buildings and Architecture',
-                        'link' => null,
-                    ],
-                    'keywords' => [
-                        0 => [
-                            'name' => 'horse',
-                        ],
-                        1 => [
-                            'name' => 'run',
-                        ],
-                        2 => [
-                            'name' => 'field',
-                        ],
-                        3 => [
-                            'name' => 'sunset',
-                        ],
-                        4 => [
-                            'name' => 'herd',
-                        ],
-                        5 => [
-                            'name' => 'wildlife',
-                        ],
-                        6 => [
-                            'name' => 'landscape',
-                        ],
-                        7 => [
-                            'name' => 'hill',
-                        ],
-                        8 => [
-                            'name' => 'sunlight',
-                        ],
-                        9 => [
-                            'name' => 'move',
-                        ],
-                        10 => [
-                            'name' => 'vitality',
-                        ],
-                        11 => [
-                            'name' => 'speed',
-                        ],
-                        12 => [
-                            'name' => 'nature',
-                        ],
-                        13 => [
-                            'name' => 'turkey',
-                        ],
-                        14 => [
-                            'name' => 'outdoors',
-                        ],
-                        15 => [
-                            'name' => 'group',
-                        ],
-                        16 => [
-                            'name' => 'animal',
-                        ],
-                        17 => [
-                            'name' => 'no people',
-                        ],
-                        18 => [
-                            'name' => 'nobody',
-                        ],
-                    ],
-                    'has_releases' => null,
-                    'asset_type_id' => null,
-                    'vector_type' => null,
-                    'content_type' => 'image/jpeg',
-                    'framerate' => null,
-                    'duration' => null,
-                    'stock_id' => null,
-                    'comps' => null,
-                    'details_url' => 'https://stock.adobe.com/299393162?as_channel=affiliate&as_source=api&as_content=7f387d439e4f4e64ae42060fc571c456',
-                    'template_type_id' => null,
-                    'template_category_ids' => null,
-                    'marketing_text' => null,
-                    'description' => null,
-                    'size_bytes' => null,
-                    'premium_level_id' => 0,
-                    'is_premium' => null,
-                    'licenses' => null,
-                    'video_preview_url' => null,
-                    'VideoPreviewHeight' => null,
-                    'video_preview_width' => null,
-                    'video_preview_content_length' => null,
-                    'video_preview_content_type' => null,
-                    'video_small_preview_url' => null,
-                    'video_small_preview_height' => null,
-                    'video_small_preview_width' => null,
-                    'video_small_preview_content_length' => null,
-                    'video_small_preview_content_type' => null,
-                ],
-                31 => [
-                    'id' => 283278803,
-                    'title' => 'Silhouette of young man with backpack in tunnel',
-                    'creator_id' => 206713618,
-                    'creator_name' => 'LIGHTFIELD STUDIOS',
-                    'creation_date' => '2019-08-11 03:29:48',
-                    'country_name' => 'United States of America',
-                    'thumbnail_url' => null,
-                    'thumbnail_html_tag' => null,
-                    'thumbnail_width' => null,
-                    'thumbnail_height' => null,
-                    'thumbnail_110_url' => null,
-                    'thumbnail_110_width' => null,
-                    'thumbnail_110_height' => null,
-                    'thumbnail_160_url' => null,
-                    'thumbnail_160_width' => null,
-                    'thumbnail_160_height' => null,
-                    'thumbnail_220_url' => null,
-                    'thumbnail_220_width' => null,
-                    'thumbnail_220_height' => null,
-                    'thumbnail_240_url' => 'https://t3.ftcdn.net/jpg/02/83/27/88/240_F_283278803_OisVnsmmW5A2JnNT7SLNMKuy7U60BMDQ.jpg',
-                    'thumbnail_240_width' => null,
-                    'thumbnail_240_height' => null,
-                    'thumbnail_500_url' => 'https://as1.ftcdn.net/jpg/02/83/27/88/500_F_283278803_OisVnsmmW5A2JnNT7SLNMKuy7U60BMDQ.jpg',
-                    'thumbnail_500_width' => null,
-                    'thumbnail_500_height' => null,
-                    'thumbnail_1000_url' => null,
-                    'thumbnail_1000_width' => null,
-                    'thumbnail_1000_height' => null,
-                    'media_type_id' => 1,
-                    'width' => 7360,
-                    'height' => 4912,
-                    'is_licensed' => null,
-                    'comp_url' => 'https://stock.adobe.com/Rest/Libraries/Watermarked/Download/283278803/2',
-                    'comp_width' => null,
-                    'comp_height' => null,
-                    'nb_views' => null,
-                    'nb_downloads' => null,
-                    'category' => [
-                        'id' => 695,
-                        'name' => 'People',
-                        'link' => null,
-                    ],
-                    'keywords' => [
-                        0 => [
-                            'name' => 'man',
-                        ],
-                        1 => [
-                            'name' => 'adult',
-                        ],
-                        2 => [
-                            'name' => 'alone',
-                        ],
-                        3 => [
-                            'name' => 'tunnel',
-                        ],
-                        4 => [
-                            'name' => 'backpack',
-                        ],
-                        5 => [
-                            'name' => 'silhouette',
-                        ],
-                        6 => [
-                            'name' => 'profile',
-                        ],
-                        7 => [
-                            'name' => 'racked',
-                        ],
-                        8 => [
-                            'name' => 'travel',
-                        ],
-                        9 => [
-                            'name' => 'dark',
-                        ],
-                        10 => [
-                            'name' => 'tour tourism',
-                        ],
-                        11 => [
-                            'name' => 'vacation',
-                        ],
-                        12 => [
-                            'name' => 'journey',
-                        ],
-                        13 => [
-                            'name' => 'trip',
-                        ],
-                        14 => [
-                            'name' => 'underpass',
-                        ],
-                        15 => [
-                            'name' => 'lifestyle',
-                        ],
-                        16 => [
-                            'name' => 'outdoors',
-                        ],
-                        17 => [
-                            'name' => 'day',
-                        ],
-                        18 => [
-                            'name' => '1',
-                        ],
-                        19 => [
-                            'name' => 'people',
-                        ],
-                    ],
-                    'has_releases' => null,
-                    'asset_type_id' => null,
-                    'vector_type' => null,
-                    'content_type' => 'image/jpeg',
-                    'framerate' => null,
-                    'duration' => null,
-                    'stock_id' => null,
-                    'comps' => null,
-                    'details_url' => 'https://stock.adobe.com/283278803?as_channel=affiliate&as_source=api&as_content=7f387d439e4f4e64ae42060fc571c456',
-                    'template_type_id' => null,
-                    'template_category_ids' => null,
-                    'marketing_text' => null,
-                    'description' => null,
-                    'size_bytes' => null,
-                    'premium_level_id' => 0,
-                    'is_premium' => null,
-                    'licenses' => null,
-                    'video_preview_url' => null,
-                    'VideoPreviewHeight' => null,
-                    'video_preview_width' => null,
-                    'video_preview_content_length' => null,
-                    'video_preview_content_type' => null,
-                    'video_small_preview_url' => null,
-                    'video_small_preview_height' => null,
-                    'video_small_preview_width' => null,
-                    'video_small_preview_content_length' => null,
-                    'video_small_preview_content_type' => null,
-                ],
-            ],
-        ];
-    }
-}
-
