+++ dev/tests/acceptance/.credentials
@@ -1,4 +1,4 @@
-magento/ADOBE_STOCK_API_KEY=%ADOBE_STOCK_API_KEY%
-magento/ADOBE_STOCK_PRIVATE_KEY=%ADOBE_STOCK_PRIVATE_KEY%
-magento/ADOBE_STOCK_USER_EMAIL=%ADOBE_STOCK_USER_EMAIL%
-magento/ADOBE_STOCK_USER_PASSWORD=%ADOBE_STOCK_USER_PASSWORD%
+magento/ADOBE_STOCK_API_KEY=adove_stock_api_key
+magento/ADOBE_STOCK_PRIVATE_KEY=adobe_stock_private_key
+magento/ADOBE_STOCK_USER_EMAIL=adobe_stock_user_email
+magento/ADOBE_STOCK_USER_PASSWORD=adobe_stock_user_password
